% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_misc.R
\name{abstractClass}
\alias{abstractClass}
\title{Declares a class as abstract.}
\usage{
abstractClass(cls, obj)
}
\arguments{
\item{cls}{The name of the abstract class to check.}

\item{obj}{The object being instantiated.}
}
\value{
Nothing.
}
\description{
Forbids instantiation of an abstract class.
This method must be called from within a constructor of an abstract class.
It will throw an error if a direct call is made to this constructor.
}
