% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_misc.R
\name{abstractMethod}
\alias{abstractMethod}
\title{Declares a method as abstract}
\usage{
abstractMethod(obj)
}
\arguments{
\item{obj}{The object on which the abstract method is called.}
}
\value{
Nothing.
}
\description{
This method must be called from within the abstract method.
}
