% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{addMsgRecObs}
\alias{addMsgRecObs}
\title{Add a message recorder to a Biodb instance.}
\usage{
addMsgRecObs(biodb)
}
\arguments{
\item{biodb}{A valid Biodb instance.}
}
\value{
The created BiodbTestMsgRec observer instance.
}
\description{
Creates a BiodbTestMsgRec observer instance and add it to a Biodb instance.
Sometimes it is required to test if a message has been emitted.
This function adds a BiodbTestMsgRec observer to a Biodb instance. This class
is target adapted to message testing, since you are then able to retrieve, for
example, the last message emitted.
}
\examples{
# Instantiate a Biodb instance for testing
biodb <- biodb::createBiodbTestInstance(log="mylogfile.log")

# Get a message recorder observer
obs <- biodb::addMsgRecObs(biodb)

# Create a connector
conn <- biodb$getFactory()$createConn('mass.csv.file')

# Delete the connector
biodb$getFactory()$deleteConn(conn)

# Get last message
obs$getLastMsg()

# Terminate the instance
biodb$terminate()

}
