% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodb.R
\docType{class}
\name{Biodb-class}
\alias{Biodb-class}
\alias{Biodb}
\title{The mother abstract class of all connection classes.}
\description{
The mother abstract class of all connection classes.
}
\section{Methods}{

\describe{
\item{\code{entriesFieldToVctOrLst(entries, field, flatten = FALSE, compute = TRUE)}}{Extract  of entries (BiodbEntry objects) into a data frame.
entries: a list containing BiodbEntry objects.
field:   
flatten: if the field has a cardinality greater than one, then values are collapsed and output is a vector of class character. 
return:  a vector or a list of values, depending on the field's type.}

\item{\code{entriesToDataframe(entries, only.atomic = TRUE, null.to.na = TRUE,
  compute = TRUE)}}{Convert a list of entries (BiodbEntry objects) into a data frame.
only.atomic Set to TRUE if you want only the atomic fields (integer, numeric, logical and character) inside the data frame.
null.to.na  If TRUE, each NULL entry gives a line of NA values inside the data frame.}
}}

