% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbBiocPersistentCache.R
\docType{class}
\name{BiodbBiocPersistentCache-class}
\alias{BiodbBiocPersistentCache-class}
\alias{BiodbBiocPersistentCache}
\title{A persistent cache implementation that uses BiocCache package.}
\description{
A persistent cache implementation that uses BiocCache package.
}

\seealso{
\code{\link{BiodbPersistentCache}},
\code{\link{BiodbBiocPersistentCache}}.
}
