% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbBiocPersistentCache.R
\name{BiodbBiocPersistentCache}
\alias{BiodbBiocPersistentCache}
\title{A persistent cache implementation that uses BiocCache package.}
\description{
A persistent cache implementation that uses BiocCache package.

A persistent cache implementation that uses BiocCache package.
}
\seealso{
\code{\link{BiodbPersistentCache}},
\code{\link{BiodbBiocPersistentCache}}.
}
\section{Super class}{
\code{\link[biodb:BiodbPersistentCache]{biodb::BiodbPersistentCache}} -> \code{BiodbBiocPersistentCache}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiodbBiocPersistentCache$new()}}
\item \href{#method-clone}{\code{BiodbBiocPersistentCache$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="deleteAllFiles">}\href{../../biodb/html/BiodbPersistentCache.html#method-deleteAllFiles}{\code{biodb::BiodbPersistentCache$deleteAllFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="deleteFile">}\href{../../biodb/html/BiodbPersistentCache.html#method-deleteFile}{\code{biodb::BiodbPersistentCache$deleteFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="deleteFiles">}\href{../../biodb/html/BiodbPersistentCache.html#method-deleteFiles}{\code{biodb::BiodbPersistentCache$deleteFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="disable">}\href{../../biodb/html/BiodbPersistentCache.html#method-disable}{\code{biodb::BiodbPersistentCache$disable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="enable">}\href{../../biodb/html/BiodbPersistentCache.html#method-enable}{\code{biodb::BiodbPersistentCache$enable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="enabled">}\href{../../biodb/html/BiodbPersistentCache.html#method-enabled}{\code{biodb::BiodbPersistentCache$enabled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="erase">}\href{../../biodb/html/BiodbPersistentCache.html#method-erase}{\code{biodb::BiodbPersistentCache$erase()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="fileExist">}\href{../../biodb/html/BiodbPersistentCache.html#method-fileExist}{\code{biodb::BiodbPersistentCache$fileExist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="fileExists">}\href{../../biodb/html/BiodbPersistentCache.html#method-fileExists}{\code{biodb::BiodbPersistentCache$fileExists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="filesExist">}\href{../../biodb/html/BiodbPersistentCache.html#method-filesExist}{\code{biodb::BiodbPersistentCache$filesExist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="folderExists">}\href{../../biodb/html/BiodbPersistentCache.html#method-folderExists}{\code{biodb::BiodbPersistentCache$folderExists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getDir">}\href{../../biodb/html/BiodbPersistentCache.html#method-getDir}{\code{biodb::BiodbPersistentCache$getDir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getFilePath">}\href{../../biodb/html/BiodbPersistentCache.html#method-getFilePath}{\code{biodb::BiodbPersistentCache$getFilePath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getFolderPath">}\href{../../biodb/html/BiodbPersistentCache.html#method-getFolderPath}{\code{biodb::BiodbPersistentCache$getFolderPath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getTmpFolderPath">}\href{../../biodb/html/BiodbPersistentCache.html#method-getTmpFolderPath}{\code{biodb::BiodbPersistentCache$getTmpFolderPath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getUsedCacheIds">}\href{../../biodb/html/BiodbPersistentCache.html#method-getUsedCacheIds}{\code{biodb::BiodbPersistentCache$getUsedCacheIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="isReadable">}\href{../../biodb/html/BiodbPersistentCache.html#method-isReadable}{\code{biodb::BiodbPersistentCache$isReadable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="isWritable">}\href{../../biodb/html/BiodbPersistentCache.html#method-isWritable}{\code{biodb::BiodbPersistentCache$isWritable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="listFiles">}\href{../../biodb/html/BiodbPersistentCache.html#method-listFiles}{\code{biodb::BiodbPersistentCache$listFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="loadFileContent">}\href{../../biodb/html/BiodbPersistentCache.html#method-loadFileContent}{\code{biodb::BiodbPersistentCache$loadFileContent()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="markerExist">}\href{../../biodb/html/BiodbPersistentCache.html#method-markerExist}{\code{biodb::BiodbPersistentCache$markerExist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="moveFilesIntoCache">}\href{../../biodb/html/BiodbPersistentCache.html#method-moveFilesIntoCache}{\code{biodb::BiodbPersistentCache$moveFilesIntoCache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="print">}\href{../../biodb/html/BiodbPersistentCache.html#method-print}{\code{biodb::BiodbPersistentCache$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="saveContentToFile">}\href{../../biodb/html/BiodbPersistentCache.html#method-saveContentToFile}{\code{biodb::BiodbPersistentCache$saveContentToFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="setMarker">}\href{../../biodb/html/BiodbPersistentCache.html#method-setMarker}{\code{biodb::BiodbPersistentCache$setMarker()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
New instance initializer. Cache objects must not be created directly.
Instead, access the cache instance through the BiodbMain instance using the
getPersistentCache() method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbBiocPersistentCache$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See the constructor of ExtGenerator for the parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbBiocPersistentCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
