% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbCircle.R
\docType{class}
\name{BiodbCircle-class}
\alias{BiodbCircle-class}
\alias{BiodbCircle}
\title{A class for representing a circle.}
\description{
This class represents a rectangle, used for graphical
representation.
}
\section{Fields}{

\describe{
\item{\code{x}}{X coordinate of center.}

\item{\code{y}}{Y coordinate of center.}

\item{\code{r}}{Radius.}
}}

\section{Methods}{

\describe{
\item{\code{getRadius()}}{Return the radius.}

\item{\code{getX()}}{Return the X coordinate.}

\item{\code{getY()}}{Return the Y coordinate.}
}}

\seealso{
\code{\link{BiodbShape}}, \code{\link{BiodbRect}}.
}
