% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbConfig.R
\docType{class}
\name{BiodbConfig-class}
\alias{BiodbConfig-class}
\alias{BiodbConfig}
\title{A class for storing configuration values.}
\description{
This class is responsible for storing configuration. You must go through the
single instance of this class to create and set and get configuration values.
To get the single instance of this class, call the \code{getConfig()} method
of class \code{Biodb}.
}
\section{Methods}{

\describe{
\item{\code{define(def)}}{Define config properties from a structured object, normally loaded from a
YAML file.}

\item{\code{disable(key)}}{Set a boolean key to FALSE.}

\item{\code{enable(key)}}{Set a boolean key to TRUE.}

\item{\code{get(key)}}{Get the value of a key.}

\item{\code{getAssocEnvVar(key)}}{Returns the environment variable associated with this configuration key.}

\item{\code{getDefaultValue(key)}}{Get the default value of a key.}

\item{\code{getDescription(key)}}{Get the description of a key.}

\item{\code{getKeys()}}{Get the list of available keys.}

\item{\code{hasKey(key)}}{Test if a key exists.}

\item{\code{isDefined(key, fail = TRUE)}}{Test if a key is defined.}

\item{\code{isEnabled(key)}}{Test if a boolean key is set to TRUE.}

\item{\code{set(key, value)}}{Set the value of a key.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get the config instance:
config <- mybiodb$getConfig()

# Print all available keys
config$getKeys()

# Get a configuration value:
value <- config$get('cache.directory')

# Set a configuration value:
config$set('cache.directory', '~/my.biodb.cache')

# For boolean values, you can use boolean methods:
config$get('offline')
config$enable('offline')    # set to TRUE
config$disable('offline')   # set to FALSE
config$isEnabled('offline')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{Biodb}}.
}
