% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbConnBase.R
\name{BiodbConnBase}
\alias{BiodbConnBase}
\title{Base class of \code{BiodbConn} for encapsulating all needed information for
database access.}
\description{
Base class of \code{BiodbConn} for encapsulating all needed information for
database access.

Base class of \code{BiodbConn} for encapsulating all needed information for
database access.
}
\details{
This is the base class for \code{BiodbConn} and \code{BiodbDbInfo}.
When defining a new connector class, your class must not inherit from
\code{BiodbBaseConn} but at least from \code{BiodbConn} (or
\code{BiodbRemoteConn} or any subclass of \code{BiodbConn}).
Its main purpose is to store property values. Those values are initialized
from YAML files. The default definition file is located inside the package
in "inst/definitions.yml" and is loaded at Biodb startup. However you can
define your own files and load them using the
\code{BiodbMain::loadDefinitions()} method.

Arguments to the contructor are:

other: Another object inheriting from \code{BiodbBaseConn}, and
from which property values will be copied.

db.class: The class of the database (\code{"mass.csv.file"},
\code{"comp.csv.file"}, ...).

properties: Some properties to set at initialization.
}
\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Accessing BiodbConnBase methods when using a BiodbDbInfo object
dbinf <- mybiodb$getDbsInfo()$get('comp.csv.file')

# Test if a property exists
dbinf$hasProp('name')

# Get a property value
dbinf$getPropertyValue('name')

# Get a property value slot
dbinf$getPropValSlot('urls', 'base.url')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Sub-classes \code{\link{BiodbDbInfo}} and \code{\link{BiodbConn}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiodbConnBase$new()}}
\item \href{#method-print}{\code{BiodbConnBase$print()}}
\item \href{#method-hasProp}{\code{BiodbConnBase$hasProp()}}
\item \href{#method-getPropSlots}{\code{BiodbConnBase$getPropSlots()}}
\item \href{#method-hasPropSlot}{\code{BiodbConnBase$hasPropSlot()}}
\item \href{#method-propExists}{\code{BiodbConnBase$propExists()}}
\item \href{#method-isSlotProp}{\code{BiodbConnBase$isSlotProp()}}
\item \href{#method-getPropValSlot}{\code{BiodbConnBase$getPropValSlot()}}
\item \href{#method-updatePropertiesDefinition}{\code{BiodbConnBase$updatePropertiesDefinition()}}
\item \href{#method-defineParsingExpressions}{\code{BiodbConnBase$defineParsingExpressions()}}
\item \href{#method-getEntryFileExt}{\code{BiodbConnBase$getEntryFileExt()}}
\item \href{#method-getDbClass}{\code{BiodbConnBase$getDbClass()}}
\item \href{#method-getConnClassName}{\code{BiodbConnBase$getConnClassName()}}
\item \href{#method-getConnClass}{\code{BiodbConnBase$getConnClass()}}
\item \href{#method-getEntryClassName}{\code{BiodbConnBase$getEntryClassName()}}
\item \href{#method-getEntryClass}{\code{BiodbConnBase$getEntryClass()}}
\item \href{#method-getEntryIdField}{\code{BiodbConnBase$getEntryIdField()}}
\item \href{#method-getPropertyValue}{\code{BiodbConnBase$getPropertyValue()}}
\item \href{#method-setPropertyValue}{\code{BiodbConnBase$setPropertyValue()}}
\item \href{#method-setPropValSlot}{\code{BiodbConnBase$setPropValSlot()}}
\item \href{#method-getBaseUrl}{\code{BiodbConnBase$getBaseUrl()}}
\item \href{#method-setBaseUrl}{\code{BiodbConnBase$setBaseUrl()}}
\item \href{#method-getWsUrl}{\code{BiodbConnBase$getWsUrl()}}
\item \href{#method-setWsUrl}{\code{BiodbConnBase$setWsUrl()}}
\item \href{#method-getToken}{\code{BiodbConnBase$getToken()}}
\item \href{#method-setToken}{\code{BiodbConnBase$setToken()}}
\item \href{#method-getName}{\code{BiodbConnBase$getName()}}
\item \href{#method-getEntryContentType}{\code{BiodbConnBase$getEntryContentType()}}
\item \href{#method-getSchedulerNParam}{\code{BiodbConnBase$getSchedulerNParam()}}
\item \href{#method-setSchedulerNParam}{\code{BiodbConnBase$setSchedulerNParam()}}
\item \href{#method-getSchedulerTParam}{\code{BiodbConnBase$getSchedulerTParam()}}
\item \href{#method-setSchedulerTParam}{\code{BiodbConnBase$setSchedulerTParam()}}
\item \href{#method-getUrls}{\code{BiodbConnBase$getUrls()}}
\item \href{#method-getUrl}{\code{BiodbConnBase$getUrl()}}
\item \href{#method-setUrl}{\code{BiodbConnBase$setUrl()}}
\item \href{#method-getXmlNs}{\code{BiodbConnBase$getXmlNs()}}
\item \href{#method-clone}{\code{BiodbConnBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$new(other = NULL, db.class = NULL, properties = NULL, cfg = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints a description of this connector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasProp"></a>}}
\if{latex}{\out{\hypertarget{method-hasProp}{}}}
\subsection{Method \code{hasProp()}}{
Tests if this connector has a property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$hasProp(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the property to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns true if the property \code{name} exists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPropSlots"></a>}}
\if{latex}{\out{\hypertarget{method-getPropSlots}{}}}
\subsection{Method \code{getPropSlots()}}{
Gets the slot fields of a property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getPropSlots(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of a property.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a character vector containing all slot names
defined.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasPropSlot"></a>}}
\if{latex}{\out{\hypertarget{method-hasPropSlot}{}}}
\subsection{Method \code{hasPropSlot()}}{
Tests if a slot property has a specific slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$hasPropSlot(name, slot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of a property.}

\item{\code{slot}}{The slot name to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns TRUE if the property \code{name} exists and has the
slot \code{slot} defined, and FALSE otherwise."
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-propExists"></a>}}
\if{latex}{\out{\hypertarget{method-propExists}{}}}
\subsection{Method \code{propExists()}}{
Checks if property exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$propExists(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of a property.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns TRUE if the property \code{name} exists, and FALSE
otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isSlotProp"></a>}}
\if{latex}{\out{\hypertarget{method-isSlotProp}{}}}
\subsection{Method \code{isSlotProp()}}{
Tests if a property is a slot property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$isSlotProp(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of a property.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns TRUE if the property is a slot propert, FALSE
otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPropValSlot"></a>}}
\if{latex}{\out{\hypertarget{method-getPropValSlot}{}}}
\subsection{Method \code{getPropValSlot()}}{
Retrieve the value of a slot of a property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getPropValSlot(name, slot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of a property.}

\item{\code{slot}}{The slot name inside the property.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the slot \code{slot} of the property \code{name}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePropertiesDefinition"></a>}}
\if{latex}{\out{\hypertarget{method-updatePropertiesDefinition}{}}}
\subsection{Method \code{updatePropertiesDefinition()}}{
Update the definition of properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$updatePropertiesDefinition(def)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{def}}{A named list of property definitions. The names of the list must
be the property names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-defineParsingExpressions"></a>}}
\if{latex}{\out{\hypertarget{method-defineParsingExpressions}{}}}
\subsection{Method \code{defineParsingExpressions()}}{
Reimplement this method in your connector class to define parsing
expressions dynamically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$defineParsingExpressions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryFileExt"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryFileExt}{}}}
\subsection{Method \code{getEntryFileExt()}}{
Returns the entry file extension used by this connector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getEntryFileExt()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character value containing the file extension.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDbClass"></a>}}
\if{latex}{\out{\hypertarget{method-getDbClass}{}}}
\subsection{Method \code{getDbClass()}}{
Gets the Biodb name of the database associated with this connector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getDbClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character value containing the Biodb database name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConnClassName"></a>}}
\if{latex}{\out{\hypertarget{method-getConnClassName}{}}}
\subsection{Method \code{getConnClassName()}}{
Gets the name of the associated connector OOP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getConnClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the connector OOP class name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConnClass"></a>}}
\if{latex}{\out{\hypertarget{method-getConnClass}{}}}
\subsection{Method \code{getConnClass()}}{
Gets the associated connector OOP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getConnClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the connector OOP class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryClassName"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryClassName}{}}}
\subsection{Method \code{getEntryClassName()}}{
Gets the name of the associated entry class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getEntryClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the name of the associated entry class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryClass"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryClass}{}}}
\subsection{Method \code{getEntryClass()}}{
Gets the associated entry class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getEntryClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the associated entry class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryIdField"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryIdField}{}}}
\subsection{Method \code{getEntryIdField()}}{
Gets the name of the corresponding database ID field in entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getEntryIdField()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the name of the database ID field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPropertyValue"></a>}}
\if{latex}{\out{\hypertarget{method-getPropertyValue}{}}}
\subsection{Method \code{getPropertyValue()}}{
Gets a property value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getPropertyValue(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the property.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the property.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPropertyValue"></a>}}
\if{latex}{\out{\hypertarget{method-setPropertyValue}{}}}
\subsection{Method \code{setPropertyValue()}}{
Sets the value of a property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setPropertyValue(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the property.}

\item{\code{value}}{The new value to set the property to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPropValSlot"></a>}}
\if{latex}{\out{\hypertarget{method-setPropValSlot}{}}}
\subsection{Method \code{setPropValSlot()}}{
Set the value of the slot of a property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setPropValSlot(name, slot, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the property.}

\item{\code{slot}}{The name of the property's slot.}

\item{\code{value}}{The new value to set the property's slot to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBaseUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getBaseUrl}{}}}
\subsection{Method \code{getBaseUrl()}}{
Returns the base URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getBaseUrl()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
THe baae URL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBaseUrl"></a>}}
\if{latex}{\out{\hypertarget{method-setBaseUrl}{}}}
\subsection{Method \code{setBaseUrl()}}{
Sets the base URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setBaseUrl(url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWsUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getWsUrl}{}}}
\subsection{Method \code{getWsUrl()}}{
Returns the web sevices URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getWsUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWsUrl"></a>}}
\if{latex}{\out{\hypertarget{method-setWsUrl}{}}}
\subsection{Method \code{setWsUrl()}}{
Sets the web sevices URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setWsUrl(ws.url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getToken"></a>}}
\if{latex}{\out{\hypertarget{method-getToken}{}}}
\subsection{Method \code{getToken()}}{
Returns the access token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getToken()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setToken"></a>}}
\if{latex}{\out{\hypertarget{method-setToken}{}}}
\subsection{Method \code{setToken()}}{
Sets the access token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setToken(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
Returns the full database name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryContentType"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryContentType}{}}}
\subsection{Method \code{getEntryContentType()}}{
Returns the entry content type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getEntryContentType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSchedulerNParam"></a>}}
\if{latex}{\out{\hypertarget{method-getSchedulerNParam}{}}}
\subsection{Method \code{getSchedulerNParam()}}{
Returns the N parameter for the scheduler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getSchedulerNParam()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSchedulerNParam"></a>}}
\if{latex}{\out{\hypertarget{method-setSchedulerNParam}{}}}
\subsection{Method \code{setSchedulerNParam()}}{
Sets the N parameter for the scheduler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setSchedulerNParam(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSchedulerTParam"></a>}}
\if{latex}{\out{\hypertarget{method-getSchedulerTParam}{}}}
\subsection{Method \code{getSchedulerTParam()}}{
Returns the T parameter for the scheduler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getSchedulerTParam()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSchedulerTParam"></a>}}
\if{latex}{\out{\hypertarget{method-setSchedulerTParam}{}}}
\subsection{Method \code{setSchedulerTParam()}}{
Sets the T parameter for the scheduler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setSchedulerTParam(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUrls"></a>}}
\if{latex}{\out{\hypertarget{method-getUrls}{}}}
\subsection{Method \code{getUrls()}}{
Returns the URLs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getUrls()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getUrl}{}}}
\subsection{Method \code{getUrl()}}{
Returns a URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getUrl(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setUrl"></a>}}
\if{latex}{\out{\hypertarget{method-setUrl}{}}}
\subsection{Method \code{setUrl()}}{
Returns a URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$setUrl(name, url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getXmlNs"></a>}}
\if{latex}{\out{\hypertarget{method-getXmlNs}{}}}
\subsection{Method \code{getXmlNs()}}{
Returns the XML namespace.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$getXmlNs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbConnBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
