% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbConnObserver.R
\docType{class}
\name{BiodbConnObserver-class}
\alias{BiodbConnObserver-class}
\alias{BiodbConnObserver}
\title{The observer class of connectors.}
\description{
Classes inheriting this class, are able to register themselves to a connector instance and receiving messages from it.
}

\seealso{
\code{\link{BiodbConn}}.
}
