% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbCsvEntry.R
\docType{class}
\name{BiodbCsvEntry-class}
\alias{BiodbCsvEntry-class}
\alias{BiodbCsvEntry}
\title{Entry class for content in CSV format.}
\description{
This is an abstract class for handling database entries whose content is in
CSV format.
}

