% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbCustomPersistentCache.R
\docType{class}
\name{BiodbCustomPersistentCache-class}
\alias{BiodbCustomPersistentCache-class}
\alias{BiodbCustomPersistentCache}
\title{A biodb custom persistent cache implementation.}
\description{
Inside the cache folder, one folder is created for each cache ID (each remote
database has one single cache ID, always the same ,except if you change its
URL). In each of theses folders are stored the cache files for this database.
}

\seealso{
\code{\link{BiodbPersistentCache}},
\code{\link{BiodbBiocPersistentCache}}.
}
