% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbDbInfo.R
\docType{class}
\name{BiodbDbInfo-class}
\alias{BiodbDbInfo-class}
\alias{BiodbDbInfo}
\title{A class for describing the characteristics of a database.}
\description{
This class is used by \code{\link{BiodbDbsInfo}} for storing database characteristics.
}
\section{Methods}{

\describe{
\item{\code{getEntryIdField()}}{:

Return the name of the corresponding database ID field in entries.}
}}

\seealso{
\code{\link{BiodbDbsInfo}}.
}
