% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbDownloadable.R
\docType{class}
\name{BiodbDownloadable-class}
\alias{BiodbDownloadable-class}
\alias{BiodbDownloadable}
\title{An abstract class (more like an interface) to model a remote database that can be downloaded locally.}
\description{
The class must be inherited by any remote database class that allows download of its whole content. The hidden/private method \code{.doDownload()} must be implemented by the database class.
}
\section{Methods}{

\describe{
\item{\code{download()}}{:

Download the database containt locally.}

\item{\code{getDownloadPath()}}{:

Get the path where the downloaded containt is written.}

\item{\code{isDownloaded()}}{:

Returns TRUE if the database containt has already been downloaded.}

\item{\code{isExtracted()}}{:

Returns TRUE of the downloaded database containt has been extracted.}
}}

\seealso{
\code{\link{RemotedbConn}}.
}
