% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbEntry.R
\docType{class}
\name{BiodbEntry-class}
\alias{BiodbEntry-class}
\alias{BiodbEntry}
\title{The mother abstract class of all database entry classes.}
\description{
The mother abstract class of all database entry classes.
}
\section{Methods}{

\describe{
\item{\code{getFieldsAsDataFrame(only.atomic = TRUE, compute = TRUE, fields = NULL)}}{Convert entry into a data frame.}

\item{\code{getFieldsAsJson(compute = TRUE)}}{Convert entry into a JSON string.}
}}

\seealso{
\code{\link{BiodbFactory}}, \code{\link{BiodbConn}}.
}
