% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbEntryField.R
\docType{class}
\name{BiodbEntryField-class}
\alias{BiodbEntryField-class}
\alias{BiodbEntryField}
\title{A class for describing an entry field.}
\description{
This class is used by \code{BiodbEntryFields} for storing field characteristics, and returning them through the \code{get()} method. The constructor is not meant to be used.
}
\section{Methods}{

\describe{
\item{\code{allowsDuplicates()}}{:

 Returns \code{TRUE} if this field allows duplicated values.}

\item{\code{hasCardMany()}}{:

 Returns \code{TRUE} if the cardinality of this field is many.}

\item{\code{hasCardOne()}}{:

 Returns \code{TRUE} if the cardinality of this field is one.}

\item{\code{isDataFrame()}}{:

 Returns \code{TRUE} if field's type is data frame.}

\item{\code{isObject()}}{:

 Returns \code{TRUE} if field's type is a class.}

\item{\code{isVector()}}{:

Returns \code{TRUE} if the field's type is one a vector type (character, integer, double or logical).}
}}

