% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbHtmlEntry.R
\docType{class}
\name{BiodbHtmlEntry-class}
\alias{BiodbHtmlEntry-class}
\alias{BiodbHtmlEntry}
\title{Entry class for content in HTML format.}
\description{
This is an abstract class for handling database entries whose content is in
HTML format.
}

\examples{
# Create a concrete entry class inheriting from this class:
MyEntry <- methods::setRefClass("MyEntry", contains="BiodbHtmlEntry")

}
\seealso{
Super class \code{\link{BiodbXmlEntry}}.
}
