% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbInfoReporter.R
\docType{class}
\name{BiodbInfoReporter-class}
\alias{BiodbInfoReporter-class}
\alias{BiodbInfoReporter}
\title{A class for display info messages for the user.}
\description{
This class is not meant to be used directly. It is automatically instantiated
inside biodb constructor in order to report information messages.
}

\seealso{
\code{\link{Biodb}}, \code{\link{BiodbObserver}}.
}
