% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbJsonEntry.R
\docType{class}
\name{BiodbJsonEntry-class}
\alias{BiodbJsonEntry-class}
\alias{BiodbJsonEntry}
\title{Entry class for content in JSON format.}
\description{
This is an abstract class for handling database entries whose content is in
JSON format.
}

\examples{
# Create a concrete entry class inheriting from CSV class:
MyEntry <- methods::setRefClass("MyEntry", contains="BiodbJsonEntry")

}
\seealso{
Super class \code{\link{BiodbEntry}}.
}
