% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbListEntry.R
\docType{class}
\name{BiodbListEntry-class}
\alias{BiodbListEntry-class}
\alias{BiodbListEntry}
\title{Entry class for content in list format.}
\description{
This is an abstract class for handling database entries whose content is in
list format.
}

\examples{
# Create a concrete entry class inheriting from CSV class:
MyEntry <- methods::setRefClass("MyEntry", contains="BiodbListEntry")

}
\seealso{
Super class \code{\link{BiodbEntry}}.
}
