% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbLogger.R
\docType{class}
\name{BiodbLogger-class}
\alias{BiodbLogger-class}
\alias{BiodbLogger}
\title{A class for logging biodb messages either to standard error or into a file.}
\description{
A class for logging biodb messages either to standard error or into a file.
}

