% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbRequest.R
\docType{class}
\name{BiodbRequest-class}
\alias{BiodbRequest-class}
\alias{BiodbRequest}
\title{Class Request.}
\arguments{
\item{url}{A \code{BiodbUrl} object.}
}
\description{
This class represents a Request object that can be used with the Request
Scheduler.
}

\seealso{
\code{\link{BiodbRequestScheduler}}, \code{\link{BiodbUrl}}.
}
