% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbRequestSchedulerRule.R
\docType{class}
\name{BiodbRequestSchedulerRule-class}
\alias{BiodbRequestSchedulerRule-class}
\alias{BiodbRequestSchedulerRule}
\title{Scheduler rule class.}
\description{
This class represents a rule for the request scheduler.
}
\details{
The constructor takes the following arguments:

host: The web host for which this rules is applicable.

n: The number of connections allowed during a period of t seconds.

t: The number of seconds during which n connections are allowed.

conn: The connector instance that is concerned by this rule.
}
\section{Methods}{

\describe{
\item{\code{addConnector(conn)}}{:

Associate a connector with this rule.
    
conn: A BiodbConn object.
    
Returned value: None.
    }

\item{\code{getConnectors()}}{:

Gets connectors associaated with this rule.
    
Returned value: A list of BiodbConn objects.
    }

\item{\code{getHost()}}{:

Gets host.
    
Returned value: Returns the host.
    }

\item{\code{getN()}}{:

Gets N value. The number of connections allowed during a period of
    T seconds.
    
Returned value: Returns N as an integer.
    }

\item{\code{getT()}}{:

Gets T value. The number of seconds during which N connections
    are allowed.
    
Returned value: Returns T as a numeric.
    }

\item{\code{removeConnector(conn)}}{:

Disassociate a connector from this rule.
    
conn: A BiodbConn instance.
    
Returned value: None.
    }

\item{\code{setFrequency(n, t)}}{:

Sets both N and T.
    
n: The number of connections allowed during a period of t seconds,
    as an integer.
    
t: The number of seconds during which n connections are allowed, as a
    numeric value.
    
Returned value: None.
    }
}}

\seealso{
\code{\link{BiodbRequestScheduler}}.
}
