% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSdfEntry.R
\docType{class}
\name{BiodbSdfEntry-class}
\alias{BiodbSdfEntry-class}
\alias{BiodbSdfEntry}
\title{Entry class for content in SDF format.}
\description{
This is an abstract class for handling database entries whose content is in
SDF format.
}

