% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSearchable.R
\docType{class}
\name{BiodbSearchable-class}
\alias{BiodbSearchable-class}
\alias{BiodbSearchable}
\title{An abstract class (more like an interface) to model a database that can be
searchable.}
\description{
The class implementing this interface, must be searchable in a generic way.
For instance, its entries must be searchable by name.
}

\seealso{
\code{\link{BiodbConn}}.
}
