% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbShape.R
\docType{class}
\name{BiodbShape-class}
\alias{BiodbShape-class}
\alias{BiodbShape}
\title{A class for representing a shape.}
\description{
This abstract class represents a shape, used for graphical representation.
}
\details{
Arguments to the constructor are:

label: A text label to associate with the shape.

color: A color, as a character string.
}
\section{Methods}{

\describe{
\item{\code{draw()}}{:

Draw the shape on the current image.
    
Returned value: None.
    }

\item{\code{getColor()}}{Return the color associated with this shape.}

\item{\code{getLabel()}}{Return the label associated with this shape.}

\item{\code{getRgbColor(alpha = 255)}}{Return the RGB color associated with this shape.}
}}

\examples{
# Create a circle instance
c <- BiodbCircle(x=12, y=5, r=3, label='MyCircle')

# Create a rectangle instance
r <- BiodbRect(left=10, top=10, bottom=20, right=30, color='yellow')

}
\seealso{
\code{\link{BiodbRect}}, \code{\link{BiodbCircle}}.
}
