% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlQuery.R
\docType{class}
\name{BiodbSqlBinaryOp-class}
\alias{BiodbSqlBinaryOp-class}
\alias{BiodbSqlBinaryOp}
\title{This class represents an SQL binary operator.}
\description{
This class represents an SQL binary operator.
}

