% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlQuery.R
\docType{class}
\name{BiodbSqlExpr-class}
\alias{BiodbSqlExpr-class}
\alias{BiodbSqlExpr}
\title{This abstract class represents an SQL expression.}
\description{
This abstract class represents an SQL expression.
}

