% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlExpr.R
\name{BiodbSqlExpr}
\alias{BiodbSqlExpr}
\title{The SQL Expression abstact class.}
\description{
The SQL Expression abstact class.

The SQL Expression abstact class.
}
\details{
This abstract class represents an SQL expression.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-toString}{\code{BiodbSqlExpr$toString()}}
\item \href{#method-clone}{\code{BiodbSqlExpr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toString"></a>}}
\if{latex}{\out{\hypertarget{method-toString}{}}}
\subsection{Method \code{toString()}}{
Converts into a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlExpr$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string containing the SQL expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlExpr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
