% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlQuery.R
\docType{class}
\name{BiodbSqlLogicalOp-class}
\alias{BiodbSqlLogicalOp-class}
\alias{BiodbSqlLogicalOp}
\title{This class represents an SQL logical operator.}
\description{
This class represents an SQL logical operator.
}

