% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlLogicalOp.R
\name{BiodbSqlLogicalOp}
\alias{BiodbSqlLogicalOp}
\title{This class represents an SQL logical operator.}
\description{
This class represents an SQL logical operator.

This class represents an SQL logical operator.
}
\section{Super class}{
\code{\link[biodb:BiodbSqlExpr]{biodb::BiodbSqlExpr}} -> \code{BiodbSqlLogicalOp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiodbSqlLogicalOp$new()}}
\item \href{#method-addExpr}{\code{BiodbSqlLogicalOp$addExpr()}}
\item \href{#method-toString}{\code{BiodbSqlLogicalOp$toString()}}
\item \href{#method-clone}{\code{BiodbSqlLogicalOp$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlLogicalOp$new(op)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{op}}{The logical operator, as a string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addExpr"></a>}}
\if{latex}{\out{\hypertarget{method-addExpr}{}}}
\subsection{Method \code{addExpr()}}{
Add an SQL expression to the logical operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlLogicalOp$addExpr(expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{A BiodbSqlExpr instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toString"></a>}}
\if{latex}{\out{\hypertarget{method-toString}{}}}
\subsection{Method \code{toString()}}{
Converts into a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlLogicalOp$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string containing the SQL expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlLogicalOp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
