% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlQuery.R
\docType{class}
\name{BiodbSqlQuery-class}
\alias{BiodbSqlQuery-class}
\alias{BiodbSqlQuery}
\title{Class SQL Query.}
\arguments{
\item{table1}{The first table of the join.}

\item{field1}{The field of the first table of the join.}

\item{table2}{The second table of the join.}

\item{field2}{The field of the second table of the join.}
}
\description{
This class represents an SQL query. It is used internally to generate an SQL query string.
}
\section{Methods}{

\describe{
\item{\code{addField(table = NULL, field)}}{:

Set the fields.}

\item{\code{addJoin(table1, field1, table2, field2)}}{:

Add a join.}

\item{\code{setDistinct(distinct)}}{:

Set or unset distinct modifier.}

\item{\code{setLimit(limit)}}{:

Set results limit.}

\item{\code{setTable(table)}}{:

Set the table.}

\item{\code{setWhere(expr)}}{:

Set  the where clause.}

\item{\code{toString()}}{:

Generate the string representation of this query.}
}}

\seealso{
\code{\link{BiodbRequestScheduler}}, \code{\link{BiodbRequest}}.
}
