% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlValue.R
\name{BiodbSqlValue}
\alias{BiodbSqlValue}
\title{This class represents an SQL value.}
\description{
This class represents an SQL value.

This class represents an SQL value.
}
\section{Super class}{
\code{biodb::BiodbSqlExpr} -> \code{BiodbSqlValue}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BiodbSqlValue-new}{\code{BiodbSqlValue$new()}}
\item \href{#method-BiodbSqlValue-toString}{\code{BiodbSqlValue$toString()}}
\item \href{#method-BiodbSqlValue-clone}{\code{BiodbSqlValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlValue-new}{}}}
\subsection{Method \code{new()}}{
Initializer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlValue$new(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlValue-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlValue-toString}{}}}
\subsection{Method \code{toString()}}{
Converts into a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlValue$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string containing the SQL expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
