% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\docType{class}
\name{BiodbTestMsgAck-class}
\alias{BiodbTestMsgAck-class}
\alias{BiodbTestMsgAck}
\title{A class for acknowledging messages during tests.}
\description{
This observer is used to call a testthat::expect_*() method each time a
message is received. This is used when running tests on Travis-CI, so Travis
does not stop tests because no change is detected in output.
}

\examples{
# To use the acknowledger, set ack=TRUE when creating the Biodb test
# instance:
biodb <- createBiodbTestInstance(ack=TRUE)

# Terminate the BiodbMain instance
biodb$terminate()

}
