% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{BiodbTestMsgAck}
\alias{BiodbTestMsgAck}
\title{A class for acknowledging messages during tests.}
\description{
This observer is used to call a testthat::expect_*() method each time a
message is received. This is used when running tests on Travis-CI, so Travis
does not stop tests because no change is detected in output.
}
\examples{
# To use the acknowledger, set ack=TRUE when creating the Biodb test
# instance:
biodb <- createBiodbTestInstance(ack=TRUE)

# Terminate the BiodbMain instance
biodb$terminate()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiodbTestMsgAck$new()}}
\item \href{#method-notifyProgress}{\code{BiodbTestMsgAck$notifyProgress()}}
\item \href{#method-clone}{\code{BiodbTestMsgAck$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbTestMsgAck$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-notifyProgress"></a>}}
\if{latex}{\out{\hypertarget{method-notifyProgress}{}}}
\subsection{Method \code{notifyProgress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbTestMsgAck$notifyProgress(what, index, total)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbTestMsgAck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
