% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\docType{class}
\name{BiodbTestMsgRec-class}
\alias{BiodbTestMsgRec-class}
\alias{BiodbTestMsgRec}
\title{A class for recording messages during tests.}
\description{
The main purpose of this class is to give access to last sent messages of the
different types: "error", "warning", "caution", "info" and "debug".
}
\section{Methods}{

\describe{
\item{\code{clearMessages()}}{:

Erase all lists of messages.
    
Returned value: None.
    }

\item{\code{getLastMsg()}}{:

Get the last message received.
    
Returned value: The last message received as a character value.
    }

\item{\code{getLastMsgByType(type)}}{:

Get the last message of a certain type.
    
type: The type of the message.
    
Returned value: The last message.
    }

\item{\code{getMsgsByType(type)}}{:

Get all messages of a certain type.
    
type: The type of the messages to retrieve.
    
Returned value: A character vector containing all messages received for
    this type.
    }

\item{\code{hasMsgs(type = NULL)}}{:

Checks if at least one message has been received.
    
type: A vector of message types on which to restrict the test.
    
Returned value: TRUE if at least one message has been received, FALSE
    otherwise.
    }

\item{\code{msg(type = "info", msg, class = NA_character_, method = NA_character_,
  lvl = 1)}}{:

Sends a message to this observer. The message will be accepted and
    handled only if the level is lower or equal to the current level for the
    message type. You can check current level for a type by calling getLevel()
    and set the level with setLevel().
    
type: The message type. It must be one of: 'info', 'debug',
    'caution', 'warning', 'error'.
    
msg: The text message to send.
    
class: The class of the object that called this message method.
    
method: The method that called this message method.
    
lvl: The level of the message.
    
Returned value: None.
    }
}}

