% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbTxtEntry.R
\docType{class}
\name{BiodbTxtEntry-class}
\alias{BiodbTxtEntry-class}
\alias{BiodbTxtEntry}
\title{Entry class for content in text format.}
\description{
This is an abstract class for handling database entries whose content is in
text format.
}

