% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbUrl.R
\docType{class}
\name{BiodbUrl-class}
\alias{BiodbUrl-class}
\alias{BiodbUrl}
\title{Class URL.}
\arguments{
\item{url}{The URL to access, as a character string.}

\item{params}{The list of parameters to use with the URL.}
}
\description{
This class represents a URL object that can be used in requests.
}

\seealso{
\code{\link{BiodbRequestScheduler}}, \code{\link{BiodbRequest}}.
}
