% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbWarningReporter.R
\docType{class}
\name{BiodbWarningReporter-class}
\alias{BiodbWarningReporter-class}
\alias{BiodbWarningReporter}
\title{A class for logging warning messages.}
\description{
This class is not meant to be used directly. It is automatically instantiate inside biodb constructor in order to report warnings.
}

\seealso{
\code{\link{Biodb}}, \code{\link{BiodbObserver}}.
}
