% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbWritable.R
\docType{class}
\name{BiodbWritable-class}
\alias{BiodbWritable-class}
\alias{BiodbWritable}
\title{An abstract class (more like an interface) to model a writable database.}
\arguments{
\item{entry}{An entry instance.}
}
\description{
A database class that implements this interface must allow the addition of new entries.
}
\section{Methods}{

\describe{
\item{\code{allowWriting()}}{Allow writing for this database.}

\item{\code{disallowWriting()}}{Disallow writing for this database.}

\item{\code{setWritingAllowed(allow)}}{Allow or disallow writing for this database.}

\item{\code{write()}}{Write the database. All modifications made to the database since the last time write() was called will be saved.}

\item{\code{writingIsAllowed()}}{Returns TRUE if writing is allowed for this database.}
}}

