% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbXmlEntry.R
\docType{class}
\name{BiodbXmlEntry-class}
\alias{BiodbXmlEntry-class}
\alias{BiodbXmlEntry}
\title{Entry class for content in XML format.}
\description{
This is an abstract class for handling database entries whose content is in
XML format.
}

\examples{
# Create a concrete entry class inheriting from CSV class:
MyEntry <- methods::setRefClass("MyEntry", contains="BiodbXmlEntry")

}
