% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChebiEntry.R
\docType{class}
\name{ChebiEntry-class}
\alias{ChebiEntry-class}
\alias{ChebiEntry}
\title{ChEBI entry class.}
\description{
This is the entry class for ChEBI database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector to ChEBI
conn <- mybiodb$getFactory()$createConn('chebi')

# Get an entry
e <- conn$getEntry('15440')

# Terminate instance.
mybiodb$terminate()

}
