% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{checkDeprecatedCacheFolders}
\alias{checkDeprecatedCacheFolders}
\title{Check deprecated default cache folders.}
\usage{
checkDeprecatedCacheFolders()
}
\value{
Nothing.
}
\description{
Searches for a deprecated location of the default cache folder, and moves
files to the new location if possible. Otherwise raises a warning.
}
\examples{
biodb::checkDeprecatedCacheFolders()

}
