% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemspiderEntry.R
\docType{class}
\name{ChemspiderEntry-class}
\alias{ChemspiderEntry-class}
\alias{ChemspiderEntry}
\title{ChemSpider entry class}
\description{
This is the entry class for the ChemSpider database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector to ChemSpider
conn <- mybiodb$getFactory()$createConn('chemspider')

# Get an entry
e <- conn$getEntry('2')

# Terminate instance.
mybiodb$terminate()

}
