% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closeMatchPpm}
\alias{closeMatchPpm}
\title{Close match PPM}
\usage{
closeMatchPpm(x, y, xidx, yidx, xolength, dppm, dmz)
}
\arguments{
\item{x}{sorted M/Z values (ascending order) of input spectrum (no NA).}

\item{y}{sorted M/Z values (ascending order) of reference spectrum (no NA).}

\item{xidx}{indices of the M/Z peaks of x, taken from the original spectrum
ordered in decreasing intensity values.}

\item{yidx}{indices of the M/Z peaks of y, taken from the original spectrum
ordered in decreasing intensity values.}

\item{xolength}{???}

\item{dppm}{???}

\item{dmz}{???}
}
\value{
???
}
\description{
Matches peaks between two spectra.
}
