% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCsvFileConn.R
\docType{class}
\name{CompCsvFileConn-class}
\alias{CompCsvFileConn-class}
\alias{CompCsvFileConn}
\title{Compound CSV File connector class.}
\description{
This is the connector class for a Compound CSV file database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Get a connector:
chebi_file <- system.file("extdata", "chebi_extract.tsv", package="biodb")
conn <- mybiodb$getFactory()$createConn('comp.csv.file', url=chebi_file)

# Get an entry
e <- conn$getEntry('')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{CsvFileConn}}.
}
