% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCsvFileEntry.R
\docType{class}
\name{CompCsvFileEntry-class}
\alias{CompCsvFileEntry-class}
\alias{CompCsvFileEntry}
\title{Compound CSV File entry class.}
\description{
This is the entry class for Compound CSV file databases.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get a connector that inherits from CsvFileConn:
chebi_file <- system.file("extdata", "chebi_extract.tsv", package="biodb")
conn <- mybiodb$getFactory()$createConn('comp.csv.file', url=chebi_file)

# Get an entry
e <- conn$getEntry('')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{BiodbCsvEntry}}.
}
