% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSqliteConn.R
\docType{class}
\name{CompSqliteConn-class}
\alias{CompSqliteConn-class}
\alias{CompSqliteConn}
\title{Class for handling a Compound database in SQLite format.}
\description{
This is the connector class for a Compound database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::BiodbMain()

# Get a connector:
chebi_file <- system.file("extdata", "chebi_extract.sqlite", package="biodb")
conn <- mybiodb$getFactory()$createConn('comp.sqlite', url=chebi_file)

# Get an entry
e <- conn$getEntry('1018')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{SqliteConn}} and interfaces
\code{\link{BiodbCompounddbConn}}.
}
