% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSqliteEntry.R
\docType{class}
\name{CompSqliteEntry-class}
\alias{CompSqliteEntry-class}
\alias{CompSqliteEntry}
\title{Compound SQLite entry class.}
\description{
This is the entry class for a Compound SQLite database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get path to LCMS database example file
lcmsdb <- system.file("extdata", "chebi_extract.sqlite", package="biodb")

# Create a connector
conn <- mybiodb$getFactory()$createConn('comp.sqlite', url=lcmsdb)

# Get an entry
e <- conn$getEntry('34.pos.col12.0.78')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{BiodbListEntry}}.
}
