% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{connNameToClassPrefix}
\alias{connNameToClassPrefix}
\title{Convert connector name into class prefix.}
\usage{
connNameToClassPrefix(connName)
}
\arguments{
\item{connName}{A connector name (e.g.: "mass.csv.file").}
}
\value{
The corresponding class prefix (e.g.: "MassCsvFile").
}
\description{
Converts the connector name into the class prefix (e.g.: "mass.csv.file" -->
"MassCsvFile").
}
