% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvFileConn.R
\name{CsvFileConn}
\alias{CsvFileConn}
\title{CSV File connector class.}
\description{
CSV File connector class.

CSV File connector class.
}
\details{
This is the abstract connector class for all CSV file databases.
}
\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Get a connector that inherits from CsvFileConn:
chebi_file <- system.file("extdata", "chebi_extract.tsv", package="biodb")
conn <- mybiodb$getFactory()$createConn('comp.csv.file', url=chebi_file)

# Get an entry
e <- conn$getEntry('1018')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super classes \code{\link{BiodbConn}},
and sub-classes \code{\link{CompCsvFileConn}},
\code{\link{MassCsvFileConn}}.
}
\section{Super classes}{
\code{\link[biodb:BiodbConnBase]{biodb::BiodbConnBase}} -> \code{\link[biodb:BiodbConn]{biodb::BiodbConn}} -> \code{CsvFileConn}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CsvFileConn$new()}}
\item \href{#method-getCsvQuote}{\code{CsvFileConn$getCsvQuote()}}
\item \href{#method-setCsvQuote}{\code{CsvFileConn$setCsvQuote()}}
\item \href{#method-getCsvSep}{\code{CsvFileConn$getCsvSep()}}
\item \href{#method-setCsvSep}{\code{CsvFileConn$setCsvSep()}}
\item \href{#method-getFieldNames}{\code{CsvFileConn$getFieldNames()}}
\item \href{#method-hasField}{\code{CsvFileConn$hasField()}}
\item \href{#method-addField}{\code{CsvFileConn$addField()}}
\item \href{#method-getFieldColName}{\code{CsvFileConn$getFieldColName()}}
\item \href{#method-setField}{\code{CsvFileConn$setField()}}
\item \href{#method-getFieldsAndColumnsAssociation}{\code{CsvFileConn$getFieldsAndColumnsAssociation()}}
\item \href{#method-getUnassociatedColumns}{\code{CsvFileConn$getUnassociatedColumns()}}
\item \href{#method-print}{\code{CsvFileConn$print()}}
\item \href{#method-setDb}{\code{CsvFileConn$setDb()}}
\item \href{#method-setIgnoreUnassignedColumns}{\code{CsvFileConn$setIgnoreUnassignedColumns()}}
\item \href{#method-clone}{\code{CsvFileConn$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getBaseUrl">}\href{../../biodb/html/BiodbConnBase.html#method-getBaseUrl}{\code{biodb::BiodbConnBase$getBaseUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getConnClass">}\href{../../biodb/html/BiodbConnBase.html#method-getConnClass}{\code{biodb::BiodbConnBase$getConnClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getConnClassName">}\href{../../biodb/html/BiodbConnBase.html#method-getConnClassName}{\code{biodb::BiodbConnBase$getConnClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getDbClass">}\href{../../biodb/html/BiodbConnBase.html#method-getDbClass}{\code{biodb::BiodbConnBase$getDbClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getEntryClass">}\href{../../biodb/html/BiodbConnBase.html#method-getEntryClass}{\code{biodb::BiodbConnBase$getEntryClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getEntryClassName">}\href{../../biodb/html/BiodbConnBase.html#method-getEntryClassName}{\code{biodb::BiodbConnBase$getEntryClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getEntryContentType">}\href{../../biodb/html/BiodbConnBase.html#method-getEntryContentType}{\code{biodb::BiodbConnBase$getEntryContentType()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getEntryFileExt">}\href{../../biodb/html/BiodbConnBase.html#method-getEntryFileExt}{\code{biodb::BiodbConnBase$getEntryFileExt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getEntryIdField">}\href{../../biodb/html/BiodbConnBase.html#method-getEntryIdField}{\code{biodb::BiodbConnBase$getEntryIdField()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getName">}\href{../../biodb/html/BiodbConnBase.html#method-getName}{\code{biodb::BiodbConnBase$getName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getPropSlots">}\href{../../biodb/html/BiodbConnBase.html#method-getPropSlots}{\code{biodb::BiodbConnBase$getPropSlots()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getPropValSlot">}\href{../../biodb/html/BiodbConnBase.html#method-getPropValSlot}{\code{biodb::BiodbConnBase$getPropValSlot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getPropertyValue">}\href{../../biodb/html/BiodbConnBase.html#method-getPropertyValue}{\code{biodb::BiodbConnBase$getPropertyValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getSchedulerNParam">}\href{../../biodb/html/BiodbConnBase.html#method-getSchedulerNParam}{\code{biodb::BiodbConnBase$getSchedulerNParam()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getSchedulerTParam">}\href{../../biodb/html/BiodbConnBase.html#method-getSchedulerTParam}{\code{biodb::BiodbConnBase$getSchedulerTParam()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getToken">}\href{../../biodb/html/BiodbConnBase.html#method-getToken}{\code{biodb::BiodbConnBase$getToken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getUrl">}\href{../../biodb/html/BiodbConnBase.html#method-getUrl}{\code{biodb::BiodbConnBase$getUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getUrls">}\href{../../biodb/html/BiodbConnBase.html#method-getUrls}{\code{biodb::BiodbConnBase$getUrls()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getWsUrl">}\href{../../biodb/html/BiodbConnBase.html#method-getWsUrl}{\code{biodb::BiodbConnBase$getWsUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="getXmlNs">}\href{../../biodb/html/BiodbConnBase.html#method-getXmlNs}{\code{biodb::BiodbConnBase$getXmlNs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="hasProp">}\href{../../biodb/html/BiodbConnBase.html#method-hasProp}{\code{biodb::BiodbConnBase$hasProp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="hasPropSlot">}\href{../../biodb/html/BiodbConnBase.html#method-hasPropSlot}{\code{biodb::BiodbConnBase$hasPropSlot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="isSlotProp">}\href{../../biodb/html/BiodbConnBase.html#method-isSlotProp}{\code{biodb::BiodbConnBase$isSlotProp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="propExists">}\href{../../biodb/html/BiodbConnBase.html#method-propExists}{\code{biodb::BiodbConnBase$propExists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setBaseUrl">}\href{../../biodb/html/BiodbConnBase.html#method-setBaseUrl}{\code{biodb::BiodbConnBase$setBaseUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setPropValSlot">}\href{../../biodb/html/BiodbConnBase.html#method-setPropValSlot}{\code{biodb::BiodbConnBase$setPropValSlot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setPropertyValue">}\href{../../biodb/html/BiodbConnBase.html#method-setPropertyValue}{\code{biodb::BiodbConnBase$setPropertyValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setSchedulerNParam">}\href{../../biodb/html/BiodbConnBase.html#method-setSchedulerNParam}{\code{biodb::BiodbConnBase$setSchedulerNParam()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setSchedulerTParam">}\href{../../biodb/html/BiodbConnBase.html#method-setSchedulerTParam}{\code{biodb::BiodbConnBase$setSchedulerTParam()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setToken">}\href{../../biodb/html/BiodbConnBase.html#method-setToken}{\code{biodb::BiodbConnBase$setToken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setUrl">}\href{../../biodb/html/BiodbConnBase.html#method-setUrl}{\code{biodb::BiodbConnBase$setUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="setWsUrl">}\href{../../biodb/html/BiodbConnBase.html#method-setWsUrl}{\code{biodb::BiodbConnBase$setWsUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConnBase" data-id="updatePropertiesDefinition">}\href{../../biodb/html/BiodbConnBase.html#method-updatePropertiesDefinition}{\code{biodb::BiodbConnBase$updatePropertiesDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="addNewEntry">}\href{../../biodb/html/BiodbConn.html#method-addNewEntry}{\code{biodb::BiodbConn$addNewEntry()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="allowEditing">}\href{../../biodb/html/BiodbConn.html#method-allowEditing}{\code{biodb::BiodbConn$allowEditing()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="allowWriting">}\href{../../biodb/html/BiodbConn.html#method-allowWriting}{\code{biodb::BiodbConn$allowWriting()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="annotateMzValues">}\href{../../biodb/html/BiodbConn.html#method-annotateMzValues}{\code{biodb::BiodbConn$annotateMzValues()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="checkDb">}\href{../../biodb/html/BiodbConn.html#method-checkDb}{\code{biodb::BiodbConn$checkDb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="collapseResultsDataFrame">}\href{../../biodb/html/BiodbConn.html#method-collapseResultsDataFrame}{\code{biodb::BiodbConn$collapseResultsDataFrame()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="correctIds">}\href{../../biodb/html/BiodbConn.html#method-correctIds}{\code{biodb::BiodbConn$correctIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="deleteAllCacheEntries">}\href{../../biodb/html/BiodbConn.html#method-deleteAllCacheEntries}{\code{biodb::BiodbConn$deleteAllCacheEntries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="deleteAllEntriesFromPersistentCache">}\href{../../biodb/html/BiodbConn.html#method-deleteAllEntriesFromPersistentCache}{\code{biodb::BiodbConn$deleteAllEntriesFromPersistentCache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="deleteAllEntriesFromVolatileCache">}\href{../../biodb/html/BiodbConn.html#method-deleteAllEntriesFromVolatileCache}{\code{biodb::BiodbConn$deleteAllEntriesFromVolatileCache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="deleteWholePersistentCache">}\href{../../biodb/html/BiodbConn.html#method-deleteWholePersistentCache}{\code{biodb::BiodbConn$deleteWholePersistentCache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="disallowEditing">}\href{../../biodb/html/BiodbConn.html#method-disallowEditing}{\code{biodb::BiodbConn$disallowEditing()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="disallowWriting">}\href{../../biodb/html/BiodbConn.html#method-disallowWriting}{\code{biodb::BiodbConn$disallowWriting()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="download">}\href{../../biodb/html/BiodbConn.html#method-download}{\code{biodb::BiodbConn$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="editingIsAllowed">}\href{../../biodb/html/BiodbConn.html#method-editingIsAllowed}{\code{biodb::BiodbConn$editingIsAllowed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="filterEntriesOnRt">}\href{../../biodb/html/BiodbConn.html#method-filterEntriesOnRt}{\code{biodb::BiodbConn$filterEntriesOnRt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getAllCacheEntries">}\href{../../biodb/html/BiodbConn.html#method-getAllCacheEntries}{\code{biodb::BiodbConn$getAllCacheEntries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getAllVolatileCacheEntries">}\href{../../biodb/html/BiodbConn.html#method-getAllVolatileCacheEntries}{\code{biodb::BiodbConn$getAllVolatileCacheEntries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getBiodb">}\href{../../biodb/html/BiodbConn.html#method-getBiodb}{\code{biodb::BiodbConn$getBiodb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getCacheFile">}\href{../../biodb/html/BiodbConn.html#method-getCacheFile}{\code{biodb::BiodbConn$getCacheFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getCacheId">}\href{../../biodb/html/BiodbConn.html#method-getCacheId}{\code{biodb::BiodbConn$getCacheId()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getChromCol">}\href{../../biodb/html/BiodbConn.html#method-getChromCol}{\code{biodb::BiodbConn$getChromCol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getDownloadPath">}\href{../../biodb/html/BiodbConn.html#method-getDownloadPath}{\code{biodb::BiodbConn$getDownloadPath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntry">}\href{../../biodb/html/BiodbConn.html#method-getEntry}{\code{biodb::BiodbConn$getEntry()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntryContent">}\href{../../biodb/html/BiodbConn.html#method-getEntryContent}{\code{biodb::BiodbConn$getEntryContent()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntryContentFromDb">}\href{../../biodb/html/BiodbConn.html#method-getEntryContentFromDb}{\code{biodb::BiodbConn$getEntryContentFromDb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntryContentRequest">}\href{../../biodb/html/BiodbConn.html#method-getEntryContentRequest}{\code{biodb::BiodbConn$getEntryContentRequest()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntryIds">}\href{../../biodb/html/BiodbConn.html#method-getEntryIds}{\code{biodb::BiodbConn$getEntryIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntryImageUrl">}\href{../../biodb/html/BiodbConn.html#method-getEntryImageUrl}{\code{biodb::BiodbConn$getEntryImageUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getEntryPageUrl">}\href{../../biodb/html/BiodbConn.html#method-getEntryPageUrl}{\code{biodb::BiodbConn$getEntryPageUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getId">}\href{../../biodb/html/BiodbConn.html#method-getId}{\code{biodb::BiodbConn$getId()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getMatchingMzField">}\href{../../biodb/html/BiodbConn.html#method-getMatchingMzField}{\code{biodb::BiodbConn$getMatchingMzField()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getMzValues">}\href{../../biodb/html/BiodbConn.html#method-getMzValues}{\code{biodb::BiodbConn$getMzValues()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getNbEntries">}\href{../../biodb/html/BiodbConn.html#method-getNbEntries}{\code{biodb::BiodbConn$getNbEntries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getNbPeaks">}\href{../../biodb/html/BiodbConn.html#method-getNbPeaks}{\code{biodb::BiodbConn$getNbPeaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="getSearchableFields">}\href{../../biodb/html/BiodbConn.html#method-getSearchableFields}{\code{biodb::BiodbConn$getSearchableFields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isCompounddb">}\href{../../biodb/html/BiodbConn.html#method-isCompounddb}{\code{biodb::BiodbConn$isCompounddb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isDownloadable">}\href{../../biodb/html/BiodbConn.html#method-isDownloadable}{\code{biodb::BiodbConn$isDownloadable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isDownloaded">}\href{../../biodb/html/BiodbConn.html#method-isDownloaded}{\code{biodb::BiodbConn$isDownloaded()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isEditable">}\href{../../biodb/html/BiodbConn.html#method-isEditable}{\code{biodb::BiodbConn$isEditable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isExtracted">}\href{../../biodb/html/BiodbConn.html#method-isExtracted}{\code{biodb::BiodbConn$isExtracted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isMassdb">}\href{../../biodb/html/BiodbConn.html#method-isMassdb}{\code{biodb::BiodbConn$isMassdb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isRemotedb">}\href{../../biodb/html/BiodbConn.html#method-isRemotedb}{\code{biodb::BiodbConn$isRemotedb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isSearchableByField">}\href{../../biodb/html/BiodbConn.html#method-isSearchableByField}{\code{biodb::BiodbConn$isSearchableByField()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="isWritable">}\href{../../biodb/html/BiodbConn.html#method-isWritable}{\code{biodb::BiodbConn$isWritable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="makeRequest">}\href{../../biodb/html/BiodbConn.html#method-makeRequest}{\code{biodb::BiodbConn$makeRequest()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="makesRefToEntry">}\href{../../biodb/html/BiodbConn.html#method-makesRefToEntry}{\code{biodb::BiodbConn$makesRefToEntry()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="msmsSearch">}\href{../../biodb/html/BiodbConn.html#method-msmsSearch}{\code{biodb::BiodbConn$msmsSearch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="requiresDownload">}\href{../../biodb/html/BiodbConn.html#method-requiresDownload}{\code{biodb::BiodbConn$requiresDownload()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchByName">}\href{../../biodb/html/BiodbConn.html#method-searchByName}{\code{biodb::BiodbConn$searchByName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchCompound">}\href{../../biodb/html/BiodbConn.html#method-searchCompound}{\code{biodb::BiodbConn$searchCompound()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchForEntries">}\href{../../biodb/html/BiodbConn.html#method-searchForEntries}{\code{biodb::BiodbConn$searchForEntries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchForMassSpectra">}\href{../../biodb/html/BiodbConn.html#method-searchForMassSpectra}{\code{biodb::BiodbConn$searchForMassSpectra()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchMsEntries">}\href{../../biodb/html/BiodbConn.html#method-searchMsEntries}{\code{biodb::BiodbConn$searchMsEntries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchMsPeaks">}\href{../../biodb/html/BiodbConn.html#method-searchMsPeaks}{\code{biodb::BiodbConn$searchMsPeaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchMzRange">}\href{../../biodb/html/BiodbConn.html#method-searchMzRange}{\code{biodb::BiodbConn$searchMzRange()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="searchMzTol">}\href{../../biodb/html/BiodbConn.html#method-searchMzTol}{\code{biodb::BiodbConn$searchMzTol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="setDownloadedFile">}\href{../../biodb/html/BiodbConn.html#method-setDownloadedFile}{\code{biodb::BiodbConn$setDownloadedFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="setEditingAllowed">}\href{../../biodb/html/BiodbConn.html#method-setEditingAllowed}{\code{biodb::BiodbConn$setEditingAllowed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="setMatchingMzField">}\href{../../biodb/html/BiodbConn.html#method-setMatchingMzField}{\code{biodb::BiodbConn$setMatchingMzField()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="setWritingAllowed">}\href{../../biodb/html/BiodbConn.html#method-setWritingAllowed}{\code{biodb::BiodbConn$setWritingAllowed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="write">}\href{../../biodb/html/BiodbConn.html#method-write}{\code{biodb::BiodbConn$write()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="BiodbConn" data-id="writingIsAllowed">}\href{../../biodb/html/BiodbConn.html#method-writingIsAllowed}{\code{biodb::BiodbConn$writingIsAllowed()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
New instance initializer. Connector classes must not be instantiated
directly. Instead, you must use the createConn() method of the factory class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{All parameters are passed to the super class initializer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCsvQuote"></a>}}
\if{latex}{\out{\hypertarget{method-getCsvQuote}{}}}
\subsection{Method \code{getCsvQuote()}}{
Gets the characters used to delimit quotes in the CSV database file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$getCsvQuote()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The characters used to delimit quotes as a single
character value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCsvQuote"></a>}}
\if{latex}{\out{\hypertarget{method-setCsvQuote}{}}}
\subsection{Method \code{setCsvQuote()}}{
Sets the characters used to delimit quotes in the CSV database file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$setCsvQuote(quote)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quote}}{The characters used to delimit quotes as a single character
value.}

\item{\code{You}}{may specify several characters. Example \"\\\"'\".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCsvSep"></a>}}
\if{latex}{\out{\hypertarget{method-getCsvSep}{}}}
\subsection{Method \code{getCsvSep()}}{
Gets the current CSV separator used for the database file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$getCsvSep()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The CSV separator as a character value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCsvSep"></a>}}
\if{latex}{\out{\hypertarget{method-setCsvSep}{}}}
\subsection{Method \code{setCsvSep()}}{
Sets the CSV separator to be used for the database file. If this method is
called after the loading of the database, it will throw an error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$setCsvSep(sep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{The CSV separator as a character value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFieldNames"></a>}}
\if{latex}{\out{\hypertarget{method-getFieldNames}{}}}
\subsection{Method \code{getFieldNames()}}{
Get the list of all biodb fields handled by this database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$getFieldNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of the biodb field names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasField"></a>}}
\if{latex}{\out{\hypertarget{method-hasField}{}}}
\subsection{Method \code{hasField()}}{
Tests if a field is defined for this database instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$hasField(field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{A valid Biodb entry field name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE of the field is defined, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addField"></a>}}
\if{latex}{\out{\hypertarget{method-addField}{}}}
\subsection{Method \code{addField()}}{
Adds a new field to the database. The field must not already exist.  The
same single value will be set to all entries for this field.  A new column
will be written in the memory data frame, containing the value given.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$addField(field, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{A valid Biodb entry field name.}

\item{\code{value}}{The value to set for this field.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFieldColName"></a>}}
\if{latex}{\out{\hypertarget{method-getFieldColName}{}}}
\subsection{Method \code{getFieldColName()}}{
Get the column name corresponding to a Biodb field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$getFieldColName(field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{A valid Biodb entry field name. This field must be defined for
this database instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The column name from the CSV file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setField"></a>}}
\if{latex}{\out{\hypertarget{method-setField}{}}}
\subsection{Method \code{setField()}}{
Sets a field by making a correspondence between a Biodb field and one or
more columns of the loaded data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$setField(field, colname, ignore.if.missing = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{A valid Biodb entry field name. This field must not be already
defined for this database instance.}

\item{\code{colname}}{A character vector containing one or more column names from
the CSV file.}

\item{\code{ignore.if.missing}}{Deprecated parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFieldsAndColumnsAssociation"></a>}}
\if{latex}{\out{\hypertarget{method-getFieldsAndColumnsAssociation}{}}}
\subsection{Method \code{getFieldsAndColumnsAssociation()}}{
Gets the association between biodb field names and CSV file column names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$getFieldsAndColumnsAssociation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with names being the biodb field names and values being a
character vector of column names from the CSV file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUnassociatedColumns"></a>}}
\if{latex}{\out{\hypertarget{method-getUnassociatedColumns}{}}}
\subsection{Method \code{getUnassociatedColumns()}}{
Gets the list of unassociated column names from the CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$getUnassociatedColumns()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector containing column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints a description of this connector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDb"></a>}}
\if{latex}{\out{\hypertarget{method-setDb}{}}}
\subsection{Method \code{setDb()}}{
Sets the database directly from a data frame. You must not have set
the database previously with the URL parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$setDb(db)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{A data frame containing your database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setIgnoreUnassignedColumns"></a>}}
\if{latex}{\out{\hypertarget{method-setIgnoreUnassignedColumns}{}}}
\subsection{Method \code{setIgnoreUnassignedColumns()}}{
Tells the connector to ignore or not the columns found in the CSV file for
which no assignment were found.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$setIgnoreUnassignedColumns(ignore)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{Set to TRUE to ignore the unassigned columns, and to FALSE
otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CsvFileConn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
