% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{error}
\alias{error}
\title{Throw an error and log it too.}
\usage{
error(...)
}
\arguments{
\item{...}{Values to be passed to sprintf().}
}
\value{
Nothing.
}
\description{
Throws am error and logs it too with biodb logger.
}
\examples{
# Throws an error:
tryCatch(biodb::error('Index is \%d.', 10), error=function(e){e$message})

}
