% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{error0}
\alias{error0}
\title{Throw an error and log it too.}
\usage{
error0(...)
}
\arguments{
\item{...}{Values to be passed to paste0().}
}
\value{
Nothing.
}
\description{
Throws an error and logs it too with biodb logger, using paste0().
}
\examples{
# Throws an error:
tryCatch(biodb::error0('Index is ', 10, '.'), error=function(e){e$message})

}
