% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpasyEnzymeConn.R
\docType{class}
\name{ExpasyEnzymeConn-class}
\alias{ExpasyEnzymeConn-class}
\alias{ExpasyEnzymeConn}
\title{The connector class to Expasy Enzyme database.}
\arguments{
\item{name}{The name to search for.}

\item{comment}{The comment to search for.}

\item{biodb.ids}{If set to \code{TRUE}, the method will return the entry IDs as a vector of characters, instead of the raw result page.}
}
\description{
This is a concrete connector class. It must never be instantiated directly, but instead be instantiated through the factory \code{\link{BiodbFactory}}. Only specific methods are described here. See super classes for the description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryContent(entry.id)}}{:

Get the content of an entry.}

\item{\code{getEntryIds(max.results = NA_integer_)}}{:

Get entry identifiers from the database.}

\item{\code{getEntryImageUrl(entry.id)}}{:

Get the URL to a picture of the entry (e.g.: a picture of the molecule in case of a compound entry).}

\item{\code{getEntryPageUrl(entry.id)}}{:

Get the URL to the page of the entry on the database web site.}

\item{\code{searchCompound(name = NULL, mass = NULL, mass.field = NULL,
  mass.tol = 0.01, mass.tol.unit = "plain", max.results = NA_integer_)}}{:

Search for compounds by name and/or by mass. For searching by mass, you must indicate a mass field to use ('monoisotopic.mass', 'molecular.mass', ...)}

\item{\code{ws.enzymeByComment(comment, biodb.ids = FALSE)}}{:

Calls enzyme-bycomment web service and returns the HTML result. See http://enzyme.expasy.org/enzyme-bycomment.html.}

\item{\code{ws.enzymeByName(name, biodb.ids = FALSE)}}{:

Calls enzyme-byname web service and returns the HTML result. See http://enzyme.expasy.org/enzyme-byname.html.}
}}

