% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpasyEnzymeEntry.R
\docType{class}
\name{ExpasyEnzymeEntry-class}
\alias{ExpasyEnzymeEntry-class}
\alias{ExpasyEnzymeEntry}
\title{Expasy Enzyme entry class.}
\description{
This is the entry class for the Expasy Enzyme database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('expasy.enzyme')

# Get an entry
e <- conn$getEntry('1.1.1.1')

# Terminate instance.
mybiodb$terminate()

}
