% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtPackage.R
\name{ExtPackage}
\alias{ExtPackage}
\title{Extension package class}
\description{
A class for generating the skeleton of a new extension package.
}
\details{
This class manages the files of an extension package.

It can generate all the files of a new extension package: DESCRIPTION,
NEWS, README.md, tests, definitons.yml, etc. Optionnaly it also generates
other files like: a \code{.travis.yml} file for Travis-CI, a Makefile for easing
development on UNIX-like platforms outside of Rstudio.

It can also upgrade files of an existing package like: definitions.yml,
Makefile, .travis.yml, LICENSE, etc.
}
\examples{
# Generate a new package:
pkgFolder <- file.path(tempfile(), 'biodbFoo')
dir.create(pkgFolder, recursive=TRUE)
biodb::ExtPackage$new(path=pkgFolder, dbName='foo.db',
                        dbTitle='Foo database', rcpp=TRUE,
                        connType='mass', entryType='txt', downloadable=TRUE,
                        remote=TRUE)$generate()

}
\section{Super class}{
\code{\link[biodb:ExtGenerator]{biodb::ExtGenerator}} -> \code{ExtPackage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExtPackage$new()}}
\item \href{#method-clone}{\code{ExtPackage$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="ExtGenerator" data-id="generate">}\href{../../biodb/html/ExtGenerator.html#method-generate}{\code{biodb::ExtGenerator$generate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodb" data-topic="ExtGenerator" data-id="upgrade">}\href{../../biodb/html/ExtGenerator.html#method-upgrade}{\code{biodb::ExtGenerator$upgrade()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtPackage$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See the constructor of ExtGenerator for the parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtPackage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
