% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_ext.R
\name{genNewExtPkg}
\alias{genNewExtPkg}
\title{Generate a new extension package for biodb.}
\usage{
genNewExtPkg(...)
}
\arguments{
\item{...}{Parameters passed to \code{\link{ExtPackage}} constructor.}
}
\value{
Nothing.
}
\description{
Generates all the necessary files for a new extension package.
}
\examples{
# Generate a new package:
pkgFolder <- file.path(tempfile(), 'biodbFoo')
dir.create(pkgFolder, recursive=TRUE)
biodb::genNewExtPkg(path=pkgFolder, dbName='foo.db',
                    dbTitle='Foo database', rcpp=TRUE,
                    connType='mass', entryType='txt', downloadable=TRUE,
                    remote=TRUE)

}
\seealso{
\code{\link{ExtPackage}}.
}
