% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{getBaseUrlContent}
\alias{getBaseUrlContent}
\title{Get URL content using base::url().}
\usage{
getBaseUrlContent(u)
}
\arguments{
\item{u}{The URL as a character value.}
}
\value{
The URL content as a character single value.
}
\description{
Get URL content using base::url().
}
