% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{getConnTypes}
\alias{getConnTypes}
\title{Get connector types.}
\usage{
getConnTypes()
}
\value{
A character vector containing the connector types.
}
\description{
Get the list of available connector types.
}
