% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{getDefaultCacheDir}
\alias{getDefaultCacheDir}
\title{Get default cache folder.}
\usage{
getDefaultCacheDir()
}
\value{
The path to the cache folder.
}
\description{
Returns the path to the default cache folder.
}
\examples{
cacheFolderPath <- biodb::getDefaultCacheDir()

}
