% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{getDefInst}
\alias{getDefInst}
\title{Get default instance of BiodbMain}
\usage{
getDefInst()
}
\value{
The default BiodbMain instance or NULL.
}
\description{
Get default instance of BiodbMain
}
\examples{
getDefInst()

}
