% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{getEntryClassName}
\alias{getEntryClassName}
\title{Get entry class name.}
\usage{
getEntryClassName(connName)
}
\arguments{
\item{connName}{A connector name (e.g.: "mass.csv.file").}
}
\value{
The name of the corresponding entry class (e.g.:
"MassCsvFileEntry").
}
\description{
Gets the name of the entry class corresponding to a connector.
}
\examples{
biodb::getEntryClassName('foo.db')

}
