% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_ext.R
\name{getLicenses}
\alias{getLicenses}
\title{Get the available licenses for extension packages.}
\usage{
getLicenses()
}
\value{
A character vector containing license names.
}
\description{
Get the available licenses for extension packages.
}
\examples{
biodb::getLicenses()

}
