% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{getLogger}
\alias{getLogger}
\title{Get the main package logger.}
\usage{
getLogger()
}
\value{
The main package logger (named "biodb") as a lgr::Logger object.
}
\description{
Gets the main package logger, parent of all loggers of this package.
}
