% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_ext.R
\name{getPkgName}
\alias{getPkgName}
\title{Get the package name from a package folder path.}
\usage{
getPkgName(pkgRoot, check = TRUE)
}
\arguments{
\item{pkgRoot}{The path to the root folder of the package.}

\item{check}{If set to TRUE the extracted package name is checked against
regular expression "^biodb[A-Z][A-Za-z0-9]+$", to ensure the format is
respected.}
}
\value{
The package name of the biodb extension.
}
\description{
The package name is extracted from the path by taking the basename.
}
