% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{getRCurlContent}
\alias{getRCurlContent}
\title{Get URL content using RCurl::getURL().}
\usage{
getRCurlContent(u, useragent = NULL)
}
\arguments{
\item{u}{The URL as a character value.}

\item{useragent}{The user agent identification.}
}
\value{
The URL content as a character single value.
}
\description{
Get URL content using RCurl::getURL().
}
