% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_ext.R
\name{getReposName}
\alias{getReposName}
\title{Extract the repository name from a package folder.}
\usage{
getReposName(pkgRoot, default = NULL)
}
\arguments{
\item{pkgRoot}{The path to the root folder of the package.}

\item{default}{A default value to return in case git4r package is not
available or the folder is not a Git repository.}
}
\value{
The repository name.
}
\description{
Given the root path of a package, returns the GitHub repository name.
}
