% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{getTestOutputDir}
\alias{getTestOutputDir}
\title{Get the test output directory.}
\usage{
getTestOutputDir()
}
\value{
The path to the test output directory, as a character value.
}
\description{
Returns the path to the test output directory. The function creates this also
this directory if it does not exist.
}
\examples{
# Get the test output directory:
biodb::getTestOutputDir()

}
