% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{getUrlContent}
\alias{getUrlContent}
\title{Get a URL content.}
\usage{
getUrlContent(u, binary = FALSE)
}
\arguments{
\item{u}{The URL as a single character value.}

\item{binary}{Set to TRUE if the content to be retrieved is binary.}
}
\value{
The content, as a single character value.
}
\description{
Get a URL content.
}
