% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmdbMetabolitesConn.R
\docType{class}
\name{HmdbMetabolitesConn-class}
\alias{HmdbMetabolitesConn-class}
\alias{HmdbMetabolitesConn}
\title{The connector class for the HMDB MetabolitesA database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{\link{BiodbFactory}}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryImageUrl(entry.id)}}{:

Gets the URL to a picture of the entry (e.g.: a picture of the
    molecule in case of a compound entry).
    
entry.id: A character vector containing entry IDs.
    
Returned value: A character vector, the same length as `entry.id`,
    containing for each entry ID either a URL or NA if no URL exists.
    }

\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }

\item{\code{getNbEntries(count = FALSE)}}{:

Get the number of entries contained in this database.
    
count: If set to TRUE and no straightforward way exists to get number of
    entries, count the output of getEntryIds().
    
Returned value: The number of entries in the database, as an integer.
    }
}}

