% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmdbMetabolitesEntry.R
\docType{class}
\name{HmdbMetabolitesEntry-class}
\alias{HmdbMetabolitesEntry-class}
\alias{HmdbMetabolitesEntry}
\title{HMDB Metabolites entry class.}
\description{
HMDB Metabolites entry class.
}

