% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{initDefInst}
\alias{initDefInst}
\title{Initialize the BiodbMain default instance}
\usage{
initDefInst(silent = FALSE)
}
\description{
Initialize the BiodbMain default instance
}
\examples{
initDefInst()

}
