% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggCompoundConn.R
\docType{class}
\name{KeggCompoundConn-class}
\alias{KeggCompoundConn-class}
\alias{KeggCompoundConn}
\title{The connector class to KEGG Compound database.}
\arguments{
\item{mass}{Single mass.}

\item{mass.min}{Minimal mass.}

\item{mass.max}{Maximal mass.}
}
\description{
This is a concrete connector class. It must never be instantiated directly, but instead be instantiated through the factory \code{\link{BiodbFactory}}. Only specific methods are described here. See super classes for the description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryImageUrl(entry.id)}}{:

Get the URL to a picture of the entry (e.g.: a picture of the molecule in case of a compound entry).}

\item{\code{searchCompound(name = NULL, mass = NULL, mass.field = NULL,
  mass.tol = 0.01, mass.tol.unit = "plain", max.results = NA_integer_)}}{:

Search for compounds by name and/or by mass. For searching by mass, you must indicate a mass field to use ('monoisotopic.mass', 'molecular.mass', ...)}

\item{\code{ws.find.exact.mass(mass = NA_real_, mass.min = NA_real_,
  mass.max = NA_real_, retfmt = c("plain", "request", "parsed", "ids"))}}{:

Search for entries by mass. See http://www.kegg.jp/kegg/docs/keggapi.html for details.}

\item{\code{ws.find.molecular.weight(mass = NA_real_, mass.min = NA_real_,
  mass.max = NA_real_, retfmt = c("plain", "request", "parsed", "ids"))}}{:

Search for entries by molecular mass. See http://www.kegg.jp/kegg/docs/keggapi.html for details.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector to KEGG Compound
conn <- mybiodb$getFactory()$createConn('kegg.compound')

# Search for compounds by exact mass
conn$ws.find.exact.mass(mass = 174.05, retfmt = 'parsed')

# Search for compounds by molecular weight 
conn$ws.find.molecular.weight(mass = 300, retfmt = 'parsed')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{KeggConn}}, \code{\link{BiodbCompounddbConn}}.
}
