% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggCompoundEntry.R
\docType{class}
\name{KeggCompoundEntry-class}
\alias{KeggCompoundEntry-class}
\alias{KeggCompoundEntry}
\title{KEGG Compound entry class.}
\description{
This is the entry class for the KEGG Compound database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.compound')

# Get an entry
e <- conn$getEntry('C00133')

# Terminate instance.
mybiodb$terminate()

}
