% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggEntry.R
\docType{class}
\name{KeggEntry-class}
\alias{KeggEntry-class}
\alias{KeggEntry}
\title{KEGG entry abstract class.}
\description{
This is the abstract entry class for all KEGG entry classes.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.compound')

# Get an entry
e <- conn$getEntry('C00133')

# Terminate instance.
mybiodb$terminate()

}
