% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggEnzymeEntry.R
\docType{class}
\name{KeggEnzymeEntry-class}
\alias{KeggEnzymeEntry-class}
\alias{KeggEnzymeEntry}
\title{KEGG Enzyme entry class.}
\description{
This is the entry class for the KEGG Enzyme class.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.enzyme')

# Get an entry
e <- conn$getEntry('1.1.1.54')

# Terminate instance.
mybiodb$terminate()

}
