% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggGenesEntry.R
\docType{class}
\name{KeggGenesEntry-class}
\alias{KeggGenesEntry-class}
\alias{KeggGenesEntry}
\title{KEGG Genes entry class.}
\description{
This is the entry class for the KEGG Genes database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.genes')

# Get an entry
e <- conn$getEntry('mmu:14635')

# Terminate instance.
mybiodb$terminate()

}
