% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggModuleConn.R
\docType{class}
\name{KeggModuleConn-class}
\alias{KeggModuleConn-class}
\alias{KeggModuleConn}
\title{The connector class to KEGG Pathway database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{\link{BiodbFactory}}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}

\seealso{
\code{\link{BiodbFactory}}, \code{\link{KeggConn}}.
}
