% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggModuleEntry.R
\docType{class}
\name{KeggModuleEntry-class}
\alias{KeggModuleEntry-class}
\alias{KeggModuleEntry}
\title{KEGG Module entry class.}
\description{
This is the entry class for KEGG Module database
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.module')

# Get an entry
e <- conn$getEntry('M00009')

# Terminate instance.
mybiodb$terminate()

}
