% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggOrthologyEntry.R
\docType{class}
\name{KeggOrthologyEntry-class}
\alias{KeggOrthologyEntry-class}
\alias{KeggOrthologyEntry}
\title{KEGG Orthology entry class.}
\description{
This is the class entry for KEGG Orthology database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.orthology')

# Get an entry
e <- conn$getEntry('K12668')

# Terminate instance.
mybiodb$terminate()

}
