% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggPathwayConn.R
\docType{class}
\name{KeggPathwayConn-class}
\alias{KeggPathwayConn-class}
\alias{KeggPathwayConn}
\title{The connector class to KEGG Pathway database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{\link{BiodbFactory}}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{buildPathwayGraph(id, directed = FALSE, drop = TRUE)}}{:

Builds a pathway graph in the form of two tables of
    vertices and edges, using KEGG database.
    
id: A character vector of KEGG pathway entry IDs.
    
directed: If set to TRUE, use available direction information
    to create directed edges, duplicating if necessary the vertices.
    
drop: If set to TRUE and the output list contains only one
    element, then the returned value is a single list of two data
    frames.
    
Returned value: A named list whose
    names are the pathway IDs, and values are lists containing two data frames
    named vertices and edges.
    }

\item{\code{convertToOrgPathways(id, org)}}{:

Takes a list of pathways IDs and converts them to the
    specified organism, filtering out the ones that do not exist in
    KEGG.
    
id: A character vector of entry IDs.
    
org: The organism in which to search for pathways, as a KEGG organism code
    (3-4 letters code, like 'hsa', 'mmu', ...). See
    https://www.genome.jp/kegg/catalog/org_list.html for a complete list of KEGG
    organism codes.
    
Returned value: A character vector, the same length as `id`,
    containing the converted IDs.
    }

\item{\code{extractPathwayMapShapes(id, color2ids)}}{:

Extracts shapes from a pathway map image.
    
id: A KEGG pathway ID.
    
color2ids: A named list defining colors for entry IDs that are present on the graph. The names of the list are standard color names. The values are character vector of entry IDs.
    
Returned value: A list of BiodbShape objects.
    }

\item{\code{getDecoratedGraphPicture(id, color2ids)}}{:

Create a pathway graph picture, with some of its elements colorized.
    
id: A KEGG pathway ID.
    
color2ids: A named list defining colors for entry IDs that are present on the graph. The names of the list are standard color names. The values are character vector of entry IDs.
    
Returned value: an image object or NULL if the package magick is not available.
    }

\item{\code{getPathwayIgraph(id, directed = FALSE, drop = TRUE)}}{:

Builds a pathway graph, as an igraph object, using KEGG database.
    
id: A character vector of KEGG pathway entry IDs.
    
directed: If set to TRUE, use available direction information
    to create directed edges, duplicating if necessary the vertices.
    
drop: If set to TRUE and the output list contains only one
    element, then the returned value is a single igraph object.
    
Returned value: A list of igraph objects, or an empty list if
    the igraph library is not available.
    }

\item{\code{getReactions(id, drop = TRUE)}}{:

Retrieves all reactions part of a KEGG pathway. Connects to
    KEGG databases, and walk through all pathways submitted, and
    their modules, to find all reactions they are composed of.
    
id: A character vector of entry IDs.
    
drop: If set to TRUE, returns a single KEGG reaction object
    instead of a list, if the list contains only one element.
    
Returned value: A list of KEGG reaction objects.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get connector
conn=mybiodb$getFactory()$createConn('kegg.pathway')

# Retrieve all reactions related to a mouse pathway:
reactions=conn$getReactions('mmu00260')

# Get a pathway graph
graph=conn$buildPathwayGraph('mmu00260')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{KeggConn}}.
}
