% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggPathwayEntry.R
\docType{class}
\name{KeggPathwayEntry-class}
\alias{KeggPathwayEntry-class}
\alias{KeggPathwayEntry}
\title{KEGG Pathway entry class.}
\description{
This is the class entry for KEGG Pathway database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.pathway')

# Get an entry
e <- conn$getEntry('map00053')

# Terminate instance.
mybiodb$terminate()

}
