% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggReactionEntry.R
\docType{class}
\name{KeggReactionEntry-class}
\alias{KeggReactionEntry-class}
\alias{KeggReactionEntry}
\title{KEGG Reaction entry class.}
\description{
This is the entry class for KEGG Reation database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.reaction')

# Get an entry
e <- conn$getEntry('R00105')

# Terminate instance.
mybiodb$terminate()

}
