% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidmapsStructureEntry.R
\docType{class}
\name{LipidmapsStructureEntry-class}
\alias{LipidmapsStructureEntry-class}
\alias{LipidmapsStructureEntry}
\title{Lipidmaps Structure entry class.}
\description{
This is the entry class for Lipidmaps Structure database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('lipidmaps.structure')

# Get an entry
e <- conn$getEntry('LMFA00000001')

# Terminate instance.
mybiodb$terminate()

}
