% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{listTestRefEntries}
\alias{listTestRefEntries}
\title{List test reference entries.}
\usage{
listTestRefEntries(conn.id)
}
\arguments{
\item{conn.id}{A valid Biodb connector ID.}
}
\value{
A list of entry IDs.
}
\description{
Lists the reference entries in the test folder for a specified connector.
}
\examples{
# List IDs of test reference entries for ChEBI:
\dontrun{
biodb::listTestRefEntries('chebi')
}

}
