% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_misc.R
\name{loadFileContents}
\alias{loadFileContents}
\title{Loads the contents of files in memory.}
\usage{
loadFileContents(x, naValues = "NA", outVect = FALSE)
}
\arguments{
\item{x}{A character vector containing the paths of the files.}

\item{naValues}{A character vector listing the content values to convert
into NA value. Set to NULL to disable the interpretation of NA values.
set to a different set of values to be interpreted.}

\item{outVect}{If set to TRUE outputs a character vector (converting any
NULL value into NA), otherwise outputs a list.}
}
\value{
A list with the contents of the files.
}
\description{
This function loads the contents of a list of files and returns the contents
as a list, each element being the content of a single file, in the same
order. If a file could not be opened, a NULL value is used as the content.
NA values are interpreted by default, but this behaviour can be turned off.
}
