% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{logDebug}
\alias{logDebug}
\title{Log debug message.}
\usage{
logDebug(...)
}
\arguments{
\item{...}{Values to be passed to sprintf().}
}
\value{
Nothing.
}
\description{
Logs a debug level message with biodb logger.
}
\examples{
# Logs a debug message:
biodb::logDebug('Index is \%d.', 10)

}
