% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{logDebug0}
\alias{logDebug0}
\title{Log debug message.}
\usage{
logDebug0(...)
}
\arguments{
\item{...}{Values to be passed to paste0()}
}
\value{
Nothing.
}
\description{
Logs a debug level message with biodb logger, using paste0().
}
\examples{
# Logs a debug message:
biodb::logDebug0('Index is ', 10, '.')

}
