% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{logInfo}
\alias{logInfo}
\title{Log information message.}
\usage{
logInfo(...)
}
\arguments{
\item{...}{Values to be passed to sprintf().}
}
\value{
Nothing.
}
\description{
Logs an information level message with biodb logger.
}
\examples{
# Logs an info message:
biodb::logInfo('Index is \%d.', 10)

}
