% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{logTrace0}
\alias{logTrace0}
\title{Log trace message.}
\usage{
logTrace0(...)
}
\arguments{
\item{...}{Values to be passed to paste0()}
}
\value{
Nothing.
}
\description{
Logs a trace level message with biodb logger, using paste0().
}
\examples{
# Logs a trace message:
biodb::logTrace0('Index is ', 10, '.')

}
