% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{lst2str}
\alias{lst2str}
\title{Convert a list into a string.}
\usage{
lst2str(x, nCut = 10)
}
\arguments{
\item{x}{The list to convert into a string.}

\item{nCut}{The maximum of elements to print.}
}
\value{
A string containing the list representation (or part of it).
}
\description{
Prints a string (partially if too big) into a string.
}
\examples{
# Converts the first 5 elements of a list into a string:
s <- lst2str(1:10, nCut=5)

}
