% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassbankConn.R
\docType{class}
\name{MassbankConn-class}
\alias{MassbankConn-class}
\alias{MassbankConn}
\title{Massbank connector class.}
\description{
This is the connector class for a Massbank database.
}
\section{Methods}{

\describe{
\item{\code{getChromCol(ids = NULL)}}{:

Gets a list of chromatographic columns contained in this database.
    
ids: A character vector of entry identifiers (i.e.: accession numbers).
    Used to restrict the set of entries on which to run the algorithm.
    
Returned value : A data.frame with two columns, one for the ID 'id' and
    another one for the title 'title'.
    }

\item{\code{getDsn(id)}}{:

Gets the DSN (database name) associated with a entry IDs.
    
id: A character vector containing entry IDs.
    
Returned value: A character vector, the same length as `id`, containing
    the corresponding DSN values, or NA values if no DSN is available.
    }

\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }

\item{\code{getNbEntries(count = FALSE)}}{:

Get the number of entries contained in this database.
    
count: If set to TRUE and no straightforward way exists to get number of
    entries, count the output of getEntryIds().
    
Returned value: The number of entries in the database, as an integer.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('massbank')

# Get the database name from a massbank id
conn$getDsn('EA256108')

# Get the URL of an entry's page
conn$getEntryPageUrl('EA256108')

# Terminate instance.
mybiodb$terminate()

}
