% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassbankEntry.R
\docType{class}
\name{MassbankEntry-class}
\alias{MassbankEntry-class}
\alias{MassbankEntry}
\title{Massbank entry class.}
\description{
This is the entry class for a Massbank database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('massbank')

# Get an entry
\dontrun{ 
e <- conn$getEntry('EA256108')
}

# Terminate instance.
mybiodb$terminate()

}
