% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassCsvFileEntry.R
\docType{class}
\name{MassCsvFileEntry-class}
\alias{MassCsvFileEntry-class}
\alias{MassCsvFileEntry}
\title{Mass CSV File entry class.}
\description{
This is the entry class for Mass CSV file databases.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get path to LCMS database example file
lcmsdb <- system.file("extdata",
                      "massbank_extract_lcms_2.tsv", package="biodb")

# Create a connector
conn <- mybiodb$getFactory()$createConn('mass.csv.file', url=lcmsdb)

# Get an entry
e <- conn$getEntry('PR010001')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{BiodbCsvEntry}}.
}
