% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSqliteEntry.R
\docType{class}
\name{MassSqliteEntry-class}
\alias{MassSqliteEntry-class}
\alias{MassSqliteEntry}
\title{Mass SQLite entry class.}
\description{
This is the entry class for a MASS SQLite database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get path to LCMS database example file
lcmsdb <- system.file("extdata", "lcmsdb.sqlite", package="biodb")

# Create a connector
conn <- mybiodb$getFactory()$createConn('mass.sqlite', url=lcmsdb)

# Get an entry
e <- conn$getEntry('34.pos.col12.0.78')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{BiodbListEntry}}.
}
