% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSqliteEntry.R
\name{MassSqliteEntry}
\alias{MassSqliteEntry}
\title{Mass spectra SQLite entry class.}
\description{
This is the entry class for a Mass spectra SQLite database.
}
\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Get path to LCMS database example file
lcmsdb <- system.file("extdata", "massbank_extract.sqlite", package="biodb")

# Create a connector
conn <- mybiodb$getFactory()$createConn('mass.sqlite', url=lcmsdb)

# Get an entry
e <- conn$getEntry('34.pos.col12.0.78')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{BiodbListEntry}}.
}
\section{Super classes}{
\code{biodb::BiodbEntry} -> \code{biodb::BiodbListEntry} -> \code{MassSqliteEntry}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MassSqliteEntry-clone}{\code{MassSqliteEntry$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="appendFieldValue"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-appendFieldValue'><code>biodb::BiodbEntry$appendFieldValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="cloneInstance"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-cloneInstance'><code>biodb::BiodbEntry$cloneInstance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="computeFields"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-computeFields'><code>biodb::BiodbEntry$computeFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="fieldHasBasicClass"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-fieldHasBasicClass'><code>biodb::BiodbEntry$fieldHasBasicClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getBiodb"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getBiodb'><code>biodb::BiodbEntry$getBiodb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getDbClass"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getDbClass'><code>biodb::BiodbEntry$getDbClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getField'><code>biodb::BiodbEntry$getField()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldCardinality"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldCardinality'><code>biodb::BiodbEntry$getFieldCardinality()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldClass"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldClass'><code>biodb::BiodbEntry$getFieldClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldDef"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldDef'><code>biodb::BiodbEntry$getFieldDef()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldNames"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldNames'><code>biodb::BiodbEntry$getFieldNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldValue"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldValue'><code>biodb::BiodbEntry$getFieldValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldsAsDataframe"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldsAsDataframe'><code>biodb::BiodbEntry$getFieldsAsDataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldsAsJson"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldsAsJson'><code>biodb::BiodbEntry$getFieldsAsJson()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldsByType"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldsByType'><code>biodb::BiodbEntry$getFieldsByType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getId"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getId'><code>biodb::BiodbEntry$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getName"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getName'><code>biodb::BiodbEntry$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getParent"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getParent'><code>biodb::BiodbEntry$getParent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="hasField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-hasField'><code>biodb::BiodbEntry$hasField()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="initialize"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-initialize'><code>biodb::BiodbEntry$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="isNew"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-isNew'><code>biodb::BiodbEntry$isNew()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="makesRefToEntry"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-makesRefToEntry'><code>biodb::BiodbEntry$makesRefToEntry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="parentIsAConnector"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-parentIsAConnector'><code>biodb::BiodbEntry$parentIsAConnector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="parseContent"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-parseContent'><code>biodb::BiodbEntry$parseContent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="print"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-print'><code>biodb::BiodbEntry$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="removeField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-removeField'><code>biodb::BiodbEntry$removeField()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="setField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-setField'><code>biodb::BiodbEntry$setField()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="setFieldValue"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-setFieldValue'><code>biodb::BiodbEntry$setFieldValue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSqliteEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MassSqliteEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSqliteEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
