% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirbaseConn.R
\docType{class}
\name{MirbaseConn-class}
\alias{MirbaseConn-class}
\alias{MirbaseConn}
\title{Mirbase connector class.}
\description{
This is the abstract connector class for Mirbase Mature connector class.
}
\section{Methods}{

\describe{
\item{\code{getNbEntries(count = FALSE)}}{:

Get the number of entries contained in this database.
    
count: If set to TRUE and no straightforward way exists to get number of
    entries, count the output of getEntryIds().
    
Returned value: The number of entries in the database, as an integer.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('mirbase.mature')

# Get an entry
\dontrun{
e <- conn$getEntry('MIMAT0000433')
}

# Terminate instance.
mybiodb$terminate()

}
