% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirbaseMatureConn.R
\docType{class}
\name{MirbaseMatureConn-class}
\alias{MirbaseMatureConn-class}
\alias{MirbaseMatureConn}
\title{Mirbase Mature connector class.}
\description{
This is the connector class for Mirbase Mature database.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }

\item{\code{searchByName(name, max.results = NA_integer_)}}{:

Searches the database for entries whose name matches the specified
    name.  Returns a character vector of entry IDs.
    
name: The name to look for.
    
max.results: If set, the number of returned IDs is limited to this
    number.
    
Returned value: A character vector of entry IDs whose name matches the
    requested name.
    }

\item{\code{wsQuery(terms, submit = "Search", retfmt = c("plain", "request", "parsed",
  "ids"))}}{:

Sends a request to the web service query.
    
terms: The query.
    
submit: Type of submission.
    
retfmt: The format to use for the returned value. 'plain' will return the
    raw results from the server, as a character value. 'request' will return the
    request a BiodbRequest object representing the request that would have been
    sent. 'parsed' will return the parsed results, as an XML object. 'ids' will
    return a character vector containing the IDs of the matching entries.
    
Returned value: Depending on `retfmt` parameter.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('mirbase.mature')

# Get an entry
e <- conn$getEntry('MIMAT0000433')

# Terminate instance.
mybiodb$terminate()

}
