% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirbaseMatureEntry.R
\docType{class}
\name{MirbaseMatureEntry-class}
\alias{MirbaseMatureEntry-class}
\alias{MirbaseMatureEntry}
\title{miRBase Mature entry class.}
\description{
This is the connector class for Mirbase Mature database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('mirbase.mature')

# Get an entry
\dontrun{
e <- conn$getEntry('MIMAT0000433')
}

# Terminate instance.
mybiodb$terminate()

}
