% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiCcdsConn.R
\docType{class}
\name{NcbiCcdsConn-class}
\alias{NcbiCcdsConn-class}
\alias{NcbiCcdsConn}
\title{NCBI CCDS connector class.}
\description{
This is the connector class for a NCBI CCDS database.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.ccds')

# Get an entry
e <- conn$getEntry('CCDS12227.1')

# Terminate instance.
mybiodb$terminate()

}
