% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiCcdsEntry.R
\docType{class}
\name{NcbiCcdsEntry-class}
\alias{NcbiCcdsEntry-class}
\alias{NcbiCcdsEntry}
\title{NCBI CCDS entry class.}
\description{
This is the entry class for a NCBI CCDS database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.ccds')

# Get an entry
e <- conn$getEntry('CCDS12227.1')

# Terminate instance.
mybiodb$terminate()

}
