% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiConn.R
\docType{class}
\name{NcbiConn-class}
\alias{NcbiConn-class}
\alias{NcbiConn}
\title{NCBI connector abstract class.}
\description{
This is an abstract class, mother of all NCBI connector classes.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.ccds')

# Get an entry
e <- conn$getEntry('CCDS12227.1')

# Terminate instance.
mybiodb$terminate()

}
