% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiEntrezConn.R
\docType{class}
\name{NcbiEntrezConn-class}
\alias{NcbiEntrezConn-class}
\alias{NcbiEntrezConn}
\title{NCBI Entrez connector abstract class.}
\description{
This is an abstract class, mother class of all NCBI Entrez connector classes.
}
\section{Methods}{

\describe{
\item{\code{getNbEntries(count = FALSE)}}{:

Get the number of entries contained in this database.
    
count: If set to TRUE and no straightforward way exists to get number of
    entries, count the output of getEntryIds().
    
Returned value: The number of entries in the database, as an integer.
    }

\item{\code{wsEfetch(id, rettype = NA_character_, retmode = NA_character_,
  retfmt = c("plain", "parsed", "request"))}}{:

Calls Entrez efetch web service. See
    https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.EFetch.
    
id: A character vector of entry IDs.
    
rettype: The retrieval type. See NCBI documentation.
    
retmode: The retrieval mode. See NCBI documentation.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed results, as an XML object. 'request' will return a BiodbRequest
    object representing the request as it would have been sent.
    
Returned value: Depending on `retfmt` parameter.
    }

\item{\code{wsEinfo(retfmt = c("plain", "request", "parsed"))}}{:

Calls Entrez einfo web service, returning information about this
    database. See https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.EInfo.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed results, as an XML object. 'request' will return a BiodbRequest
    object representing the request as it would have been sent.
    
Returned value: Depending on `retfmt` parameter.
    }

\item{\code{wsEsearch(term, field = NA_character_, retmax = NA_integer_,
  retfmt = c("plain", "parsed", "request", "ids"))}}{:

Calls Entrez esearch web service. See
    https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch.
    
term: Text query. See NCBI documentation.
    
field: Entrez field to which to limit the search. See NCBI documentation.
    
retmax: Maximum number of entry IDs to return.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed results, as an XML object. 'request' will return a BiodbRequest
    object representing the request as it would have been sent. 'ids' will
    return a character vector containing the IDs of the matching entries.
    
Returned value: Depending on `retfmt` parameter.
    }
}}

