% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiGeneConn.R
\docType{class}
\name{NcbiGeneConn-class}
\alias{NcbiGeneConn-class}
\alias{NcbiGeneConn}
\title{NCBI Gene connector class.}
\description{
This is the connector class for a NCBI Gene database.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }

\item{\code{searchByName(name, max.results = NA_integer_)}}{:

Searches the database for entries whose name matches the specified
    name.  Returns a character vector of entry IDs.
    
name: The name to look for.
    
max.results: If set, the number of returned IDs is limited to this
    number.
    
Returned value: A character vector of entry IDs whose name matches the
    requested name.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.gene')

# Get an entry
e <- conn$getEntry('2833')

# Terminate instance.
mybiodb$terminate()

}
