% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiGeneEntry.R
\docType{class}
\name{NcbiGeneEntry-class}
\alias{NcbiGeneEntry-class}
\alias{NcbiGeneEntry}
\title{NCBI Gene entry class.}
\description{
This is the entry class for a NCBI Gene database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.gene')

# Get an entry
e <- conn$getEntry('2833')

# Terminate instance.
mybiodb$terminate()

}
