% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemCompEntry.R
\docType{class}
\name{NcbiPubchemCompEntry-class}
\alias{NcbiPubchemCompEntry-class}
\alias{NcbiPubchemCompEntry}
\title{NCBI PubChem Compound entry class.}
\description{
NCBI PubChem Compound entry class.
}

