% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemEntry.R
\docType{class}
\name{NcbiPubchemEntry-class}
\alias{NcbiPubchemEntry-class}
\alias{NcbiPubchemEntry}
\title{NCBI PubChem entry class.}
\description{
This the abstract entry class for all NCBI PubChem entry classes.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.pubchem.comp')

# Get an entry
e <- conn$getEntry('2')

# Terminate instance.
mybiodb$terminate()

}
