% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemSubstConn.R
\docType{class}
\name{NcbiPubchemSubstConn-class}
\alias{NcbiPubchemSubstConn-class}
\alias{NcbiPubchemSubstConn}
\title{NCBI PubChem Substance connector class.}
\description{
This is the connector class for a NCBI PubChen Substance database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.pubchem.subst')

# Get an entry
e <- conn$getEntry('2')

# Terminate instance.
mybiodb$terminate()

}
