% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemSubstEntry.R
\docType{class}
\name{NcbiPubchemSubstEntry-class}
\alias{NcbiPubchemSubstEntry-class}
\alias{NcbiPubchemSubstEntry}
\title{NCBI PubChem Substance entry class.}
\description{
This is the entry class for a NCBI PubChen Substance database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.pubchem.subst')

# Get an entry
e <- conn$getEntry('2')

# Terminate instance.
mybiodb$terminate()

}
