% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NciCactusEntry.R
\docType{class}
\name{NciCactusEntry-class}
\alias{NciCactusEntry-class}
\alias{NciCactusEntry}
\title{NCI CACTUS entry class.}
\description{
This is the entry class for the NCI CACTUS database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('nci.cactus')

# Get an entry
e <- conn$getEntry('749674')

# Terminate instance.
mybiodb$terminate()

}
