% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{newInst}
\alias{newInst}
\title{Create a new BiodbMain instance.}
\usage{
newInst(...)
}
\arguments{
\item{...}{The parameters to pass to the BiodbMain constructor. See
\code{\link{BiodbMain}}.}
}
\value{
A new BiodbMain instance.
}
\description{
Instantiates a new BiodbMain object by calling the constructor.
}
\examples{
# Create a new BiodbMain instance:
mybiodb <- biodb::newInst()

# Terminate the instance:
mybiodb$terminate()
}
\seealso{
\code{\link{BiodbMain}}.
}
