% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakforestCompoundEntry.R
\docType{class}
\name{PeakforestCompoundEntry-class}
\alias{PeakforestCompoundEntry-class}
\alias{PeakforestCompoundEntry}
\title{PeakForest Compound entry class.}
\description{
PeakForest Compound entry class.
}

