% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakforestConn.R
\docType{class}
\name{PeakforestConn-class}
\alias{PeakforestConn-class}
\alias{PeakforestConn}
\title{PeakForest connector abstract class.}
\description{
This abstract class is the mother class of all PeakForest connector classes.
}
\section{Methods}{

\describe{
\item{\code{getNbEntries(count = FALSE)}}{:

Get the number of entries contained in this database.
    
count: If set to TRUE and no straightforward way exists to get number of
    entries, count the output of getEntryIds().
    
Returned value: The number of entries in the database, as an integer.
    }

\item{\code{wsAllCount(retfmt = c("plain", "request", "parsed"))}}{:

Calls the all/count web service, and return the number of entries
    contained inside the database.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed result, as an integer value. 'request' will return a BiodbRequest
    object representing the request as it would have been sent.
    
Returned value: Depending on `retfmt` parameter.
    }

\item{\code{wsAllIds(retfmt = c("plain", "request", "parsed", "ids"))}}{:

Calls the all/ids web service, and return full list of entry IDs
    contained inside the database.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed result, as a JSON object. 'request' will return a BiodbRequest
    object representing the request as it would have been sent. 'ids' will
    return a character vector containing the entry IDs.
    
Returned value: Depending on `retfmt` parameter.
    }

\item{\code{wsSearch(term, max = NA_integer_, retfmt = c("plain", "request", "parsed",
  "ids"))}}{:

Calls the search web service.
    
term: The text to search for.
    
max: The maximum number of matching entries to return.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed results, as a JSON object. 'request' will return a BiodbRequest
    object representing the request as it would have been sent. 'ids' will
    return a character vector containing the IDs of the matching entries.
    
Returned value: Depending on `retfmt` parameter.
    }
}}

