% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakforestMassEntry.R
\docType{class}
\name{PeakforestMassEntry-class}
\alias{PeakforestMassEntry-class}
\alias{PeakforestMassEntry}
\title{PeakForest Mass entry class.}
\description{
This is the entry class for PeakForest Mass database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('peakforest.mass')

# Get an entry
e <- conn$getEntry('1000')

# Terminate instance.
mybiodb$terminate()

}
