% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_misc.R
\name{prepareFileContents}
\alias{prepareFileContents}
\title{Prepares file contents for saving.}
\usage{
prepareFileContents(contents)
}
\arguments{
\item{contents}{File contents, as a list or a character vector.}
}
\value{
File contents.
}
\description{
Prepares file contents for saving.
}
