% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_misc.R
\name{saveContentsToFiles}
\alias{saveContentsToFiles}
\title{Saves contents to files.}
\usage{
saveContentsToFiles(files, contents, prepareContents = FALSE)
}
\arguments{
\item{files}{The file paths to use for saving contents.}

\item{contents}{The contents to save, as a list or a character vector.}

\item{prepareContents}{If set to TRUE, then calls prepareFileContents() on
the contents before saving them.}
}
\value{
Nothing.
}
\description{
Saves contents to files.
}
