% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{setTestContext}
\alias{setTestContext}
\title{Set a test context.}
\usage{
setTestContext(biodb, text)
}
\arguments{
\item{biodb}{A valid Biodb instance.}

\item{text}{The text to print as test context.}
}
\value{
No value returned.
}
\description{
Define a context for tests using testthat framework.
In addition to calling `testthat::context()`, the function will call
`biodb::Biodb::info()` to signal the context to observers. This will allow
to print the test context also into the test log file, if logger is used
for tests.
}
\examples{
# Define a context before running tests:
\dontrun{
setTestContext(biodb, "Test my database connector.")
}

}
