% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqliteConn.R
\docType{class}
\name{SqliteConn-class}
\alias{SqliteConn-class}
\alias{SqliteConn}
\title{SQLite connector class.}
\description{
This is the abstract connector class for all SQLite databases.
}
\section{Methods}{

\describe{
\item{\code{getEntryContentFromDb(entry.id)}}{:

Get the contents of entries directly from the database. A direct
    request or an access to the database will be made in order to retrieve the
    contents. No access to the biodb cache system will be made.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A character vector, the same size of entry.id, with
    contents of the requested entries. An NA value will be set for the content
    of each entry for which the retrieval failed.
    }

\item{\code{isSearchableByField(field)}}{:

Tests if a field can be used to search entries when using methods
    searchByName() and searchCompound().
    
field: The name of the field.
    
Returned value: Returns TRUE if the database is searchable using the
    specified field, FALSE otherwise.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Get a connector that inherits from SqliteConn:
chebi_file <- system.file("extdata", "chebi_extract.sqlite", package="biodb")
conn <- mybiodb$getFactory()$createConn('comp.sqlite', url=chebi_file)

# Get an entry
e <- conn$getEntry('1018')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super classes \code{\link{BiodbConn}}, \code{\link{BiodbWritable}},
\code{\link{BiodbEditable}}, and sub-classes \code{\link{CompSqliteConn}},
and \code{\link{MassSqliteConn}}.
}
