% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{testContext}
\alias{testContext}
\title{Set a test context.}
\usage{
testContext(text)
}
\arguments{
\item{text}{The text to print as test context.}
}
\value{
No value returned.
}
\description{
Define a context for tests using testthat framework.
In addition to calling \code{testthat::context()}, the function will call
\verb{biodb::Biodb::info()} to signal the context to observers. This will allow
to print the test context also into the test log file, if logger is used
for tests.
}
\examples{
# Define a context before running tests:
biodb::testContext("Test my database connector.")

# Instantiate a Biodb instance for testing
biodb <- biodb::createBiodbTestInstance()

# Terminate the instance
biodb$terminate()

}
