% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotEntry.R
\docType{class}
\name{UniprotEntry-class}
\alias{UniprotEntry-class}
\alias{UniprotEntry}
\title{Uniprot entry class.}
\description{
This is the entry class for Uniprot database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('uniprot')

# Get an entry
e <- conn$getEntry('P01011')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super class \code{\link{BiodbXmlEntry}}.
}
