% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrlRequestScheduler.R
\docType{class}
\name{UrlRequestScheduler-class}
\alias{UrlRequestScheduler-class}
\alias{UrlRequestScheduler}
\title{Class for handling URL requests.}
\arguments{
\item{url}{The URL to access, as a character string.}

\item{soap.request}{The XML SOAP request to send, as a character string.}

\item{soap.action}{The SOAP action to contact, as a character string.}

\item{params}{The list of parameters to use with the URL.}

\item{method}{The method to use: either 'get' or 'post'.}

\item{opts}{The CURL options to use.}

\item{dest.file}{A path to a destination file.}
}
\description{
This class handles GET and POST requests, as well as file downloading. Each remote database connection instance (instance of concrete class inheriting from \code{RemotedbConn}) creates an instance of \code{UrlRequestScheduler} for handling database connection. A timer is used to schedule connections, and avoid sending too much requests to the database. This class is not meant to be used directly by the library user. See section Fields for a list of the constructor's parameters.
}
\section{Fields}{

\describe{
\item{\code{n}}{The number of connections allowed for each t seconds.}

\item{\code{t}}{The number of seconds during which n connections are allowed.}
}}

\section{Methods}{

\describe{
\item{\code{downloadFile(url, dest.file)}}{:

Download the content of a URL and save it into the specified destination file.}

\item{\code{getUrl(url, params = list(), method = "get",
  opts = .self$.get.curl.opts())}}{:

Send a URL request, either with GET or POST method, and return result.}

\item{\code{sendSoapRequest(url, soap.request, soap.action = NA_character_)}}{:

Send a SOAP request to a URL. Returns the string result.}
}}

\seealso{
\code{\link{RemotedbConn}}.
}
