% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{warn}
\alias{warn}
\title{Throw a warning and log it too.}
\usage{
warn(...)
}
\arguments{
\item{...}{Values to be passed to sprintf().}
}
\value{
Nothing.
}
\description{
Throws a warning and logs it too with biodb logger.
}
\examples{
# Throws a warning:
tryCatch(biodb::warn('Index is \%d.', 10), warning=function(w){w$message})

}
