% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbBiocPersistentCache.R
\name{BiodbBiocPersistentCache}
\alias{BiodbBiocPersistentCache}
\title{A persistent cache implementation that uses BiocCache package.}
\description{
A persistent cache implementation that uses BiocCache package.

A persistent cache implementation that uses BiocCache package.
}
\seealso{
\code{\link{BiodbPersistentCache}},
\code{\link{BiodbBiocPersistentCache}}.
}
\section{Super class}{
\code{\link[biodb:BiodbPersistentCache]{biodb::BiodbPersistentCache}} -> \code{BiodbBiocPersistentCache}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BiodbBiocPersistentCache-new}{\code{BiodbBiocPersistentCache$new()}}
\item \href{#method-BiodbBiocPersistentCache-clone}{\code{BiodbBiocPersistentCache$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="addFilesToCache"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-addFilesToCache'><code>biodb::BiodbPersistentCache$addFilesToCache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="copyFilesIntoCache"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-copyFilesIntoCache'><code>biodb::BiodbPersistentCache$copyFilesIntoCache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="deleteAllFiles"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-deleteAllFiles'><code>biodb::BiodbPersistentCache$deleteAllFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="deleteFile"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-deleteFile'><code>biodb::BiodbPersistentCache$deleteFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="deleteFiles"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-deleteFiles'><code>biodb::BiodbPersistentCache$deleteFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="disable"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-disable'><code>biodb::BiodbPersistentCache$disable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="enable"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-enable'><code>biodb::BiodbPersistentCache$enable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="enabled"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-enabled'><code>biodb::BiodbPersistentCache$enabled()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="erase"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-erase'><code>biodb::BiodbPersistentCache$erase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="fileExist"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-fileExist'><code>biodb::BiodbPersistentCache$fileExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="fileExists"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-fileExists'><code>biodb::BiodbPersistentCache$fileExists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="filesExist"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-filesExist'><code>biodb::BiodbPersistentCache$filesExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="folderExists"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-folderExists'><code>biodb::BiodbPersistentCache$folderExists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getDir"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-getDir'><code>biodb::BiodbPersistentCache$getDir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getFilePath"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-getFilePath'><code>biodb::BiodbPersistentCache$getFilePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getFolderPath"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-getFolderPath'><code>biodb::BiodbPersistentCache$getFolderPath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getTmpFolderPath"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-getTmpFolderPath'><code>biodb::BiodbPersistentCache$getTmpFolderPath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="getUsedCacheIds"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-getUsedCacheIds'><code>biodb::BiodbPersistentCache$getUsedCacheIds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="isReadable"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-isReadable'><code>biodb::BiodbPersistentCache$isReadable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="isWritable"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-isWritable'><code>biodb::BiodbPersistentCache$isWritable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="listFiles"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-listFiles'><code>biodb::BiodbPersistentCache$listFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="loadFileContent"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-loadFileContent'><code>biodb::BiodbPersistentCache$loadFileContent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="markerExist"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-markerExist'><code>biodb::BiodbPersistentCache$markerExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="markerExists"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-markerExists'><code>biodb::BiodbPersistentCache$markerExists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="moveFilesIntoCache"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-moveFilesIntoCache'><code>biodb::BiodbPersistentCache$moveFilesIntoCache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="print"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-print'><code>biodb::BiodbPersistentCache$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="saveContentToFile"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-saveContentToFile'><code>biodb::BiodbPersistentCache$saveContentToFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbPersistentCache" data-id="setMarker"><a href='../../biodb/html/BiodbPersistentCache.html#method-BiodbPersistentCache-setMarker'><code>biodb::BiodbPersistentCache$setMarker()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbBiocPersistentCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbBiocPersistentCache-new}{}}}
\subsection{Method \code{new()}}{
New instance initializer. Cache objects must not be created directly.
Instead, access the cache instance through the BiodbMain instance using the
getPersistentCache() method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbBiocPersistentCache$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See the constructor of ExtGenerator for the parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbBiocPersistentCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbBiocPersistentCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbBiocPersistentCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
