% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbCircle.R
\docType{class}
\name{BiodbCircle-class}
\alias{BiodbCircle-class}
\alias{BiodbCircle}
\title{A class for representing a circle.}
\description{
This class represents a rectangle, used for graphical
representation. It is used by
KeggPathwayConn::extractPathwayMapShapes() method.
}
\details{
Arguments to the constructor are:

x: X coordinate of center.

y: Y coordinate of center.

r: Radius.
}
\section{Methods}{

\describe{
\item{\code{draw()}}{:

Draw the shape on the current image.
    
Returned value: None.
    }

\item{\code{getRadius()}}{:

Get the radius.
    
Returned value: The radius.
    }

\item{\code{getX()}}{:

Get the X coordinate.
    
Returned value: The X coordinate.
    }

\item{\code{getY()}}{:

Get the Y coordinate.
    
Returned value: The Y coordinate.
    }
}}

\examples{
# Create an instance
c1 <- BiodbCircle(x=12, y=5, r=3)

# Since it inherits from BiodbShape, a color and a label can be set
c2 <- BiodbCircle(x=12, y=5, r=3, color='blue', label='Circle 2')

# Getting center
c1$getX()
c1$getY()

# Getting radius
c1#getRadius()

# Draw a circle on the current image
\dontrun{
c1$draw()
}

}
\seealso{
\code{\link{BiodbShape}}, \code{\link{BiodbRect}}.
}
