% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbCircle.R
\docType{class}
\name{BiodbCircle-class}
\alias{BiodbCircle-class}
\alias{BiodbCircle}
\title{A class for representing a circle.}
\description{
This class represents a rectangle, used for graphical
representation. It is used inside by
KeggPathwayConn::extractPathwayMapShapes() method.
}
\section{Fields}{

\describe{
\item{\code{x}}{X coordinate of center.}

\item{\code{y}}{Y coordinate of center.}

\item{\code{r}}{Radius.}
}}

\section{Methods}{

\describe{
\item{\code{getRadius()}}{Return the radius.}

\item{\code{getX()}}{Return the X coordinate.}

\item{\code{getY()}}{Return the Y coordinate.}
}}

\examples{
# Create an instance
c1 <- BiodbCircle(x=12, y=5, r=3)

# Since it inherits from BiodbShape, a color and a label can be set
c2 <- BiodbCircle(x=12, y=5, r=3, color='blue', label='Circle 2')

# Getting center
c1$getX()
c1$getY()

# Getting radius
c1#getRadius()

# Draw a circle on the current image
\dontrun{
c1$draw()
}

}
\seealso{
\code{\link{BiodbShape}}, \code{\link{BiodbRect}}.
}
