% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbCompounddbConn.R
\docType{class}
\name{BiodbCompounddbConn-class}
\alias{BiodbCompounddbConn-class}
\alias{BiodbCompounddbConn}
\title{An interface for all compound databases.}
\arguments{
\item{name}{The name of a compound.}

\item{mass}{The searched mass.}

\item{mass.field}{The mass field to use for the search. For instance: 'molecular.mass', 'monoisotopic.mass', 'average.mass'.}

\item{mass.tol}{The tolerance on the molecular mass.}

\item{mass.tol.unit}{The unit used for molecular mass tolerance. Either 'plain' or 'ppm'.}

\item{max.results}{The maximum number of matches wanted.}
}
\description{
This interface must be inherited by all compound databases. It declares method headers specific to compound databases.
}
\section{Methods}{

\describe{
\item{\code{searchCompound(name = NULL, mass = NULL, mass.field = NULL,
  mass.tol = 0.01, mass.tol.unit = "plain", max.results = NA_integer_)}}{:

Search for compounds by name and/or by mass. For searching by mass, you must indicate a mass field to use ('monoisotopic.mass', 'molecular.mass', ...)}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get the connector of a compound database
conn <- mybiodb$getFactory()$createConn('chebi')

# Search for compounds
conn$searchCompound(name = 'prion protein', max.results = 10)

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbConn}}.
}
