% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbConfig.R
\docType{class}
\name{BiodbConfig-class}
\alias{BiodbConfig-class}
\alias{BiodbConfig}
\title{A class for storing configuration values.}
\description{
This class is responsible for storing configuration. You must go through the single instance of this class to create and set and get configuration values. To get the single instance of this class, call the \code{getConfig()} method of class \code{Biodb}.
}
\section{Methods}{

\describe{
\item{\code{disable(key)}}{:

Set a boolean key to FALSE.}

\item{\code{enable(key)}}{:

Set a boolean key to TRUE.}

\item{\code{get(key)}}{:

Get a the value of a key.}

\item{\code{getDefaultValue(key)}}{:

Get the default value of a key.}

\item{\code{getDescription(key)}}{:

Get the description of a key.}

\item{\code{getKeys()}}{:

Get the list of available keys.}

\item{\code{isDefined(key)}}{:

Test if a key is defined.}

\item{\code{isEnabled(key)}}{:

Test if a boolean key is set to TRUE.}

\item{\code{set(key, value)}}{:

Set a the value of a key.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get the config instance:
config <- mybiodb$getConfig()

# Print all available keys
print(config$getKeys())

# Get a configuration value:
value <- config$get('cache.directory')

# Set a configuration value:
config$set('cache.directory', '~/my.biodb.cache')

# For boolean values, you can use boolean methods:
print(config$get('offline'))
config$enable('offline')    # set to TRUE
config$disable('offline')   # set to FALSE
print(if (config$isEnabled('offline')) 'Mode offline is ON.' else 'Mode offline is OFF.')

}
\seealso{
\code{\link{Biodb}}.
}
