% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbConnBase.R
\docType{class}
\name{BiodbConnBase-class}
\alias{BiodbConnBase-class}
\alias{BiodbConnBase}
\title{Base class of \code{BiodbConn} for encapsulating all needed information for database access.}
\arguments{
\item{base.url}{A base URL for the database.}

\item{n}{The scheduler "number of requests parameter.}

\item{t}{The scheduler "time" parameter, in seconds.}

\item{token}{An access token for the database.}

\item{ws.url}{A web service URL for the database.}

\item{xml.ns}{The XML namespace used by the database.}
}
\description{
This is the base class for \code{BiodbConn} and \code{BiodbDbInfo}. The constructor is not meant to be used, but for development purposes the constructor's parameters are nevertheless described in the Fields section.
}
\section{Fields}{

\describe{
\item{\code{db.class}}{The class of the database (\code{"massbank", "hmdb.metabolies", ...}).}

\item{\code{base.url}}{The main URL of the database.}

\item{\code{ws.url}}{The web services URL of the database.}

\item{\code{xml.ns}}{The XML namespace used by the database.}

\item{\code{token}}{An access token for the database.}

\item{\code{scheduler.n}}{The N parameter for the scheduler. The N paramter is the number of request allowed for each bit of time (defined by the T parameter).}

\item{\code{scheduler.t}}{The T parameter for the scheduler. The bit of time, in seconds, during which a maximum of N (see scheduler.n) requests is allowed.}
}}

\section{Methods}{

\describe{
\item{\code{getBaseUrl()}}{:

Returns the base URL.}

\item{\code{getConnClass()}}{:

Returns the associated connection class.}

\item{\code{getConnClassName()}}{:

Returns the name of the associated connection class.}

\item{\code{getEntryClass()}}{:

Returns the associated entry class.}

\item{\code{getEntryClassName()}}{:

Returns the name of the associated entry class.}

\item{\code{getEntryContentType()}}{:

Returns the entry content type.}

\item{\code{getEntryIdField()}}{:

Return the name of the corresponding database ID field in entries.}

\item{\code{getName()}}{:

Returns the full database name.}

\item{\code{getSchedulerNParam()}}{:

Returns the N parameter for the scheduler.}

\item{\code{getSchedulerTParam()}}{:

Returns the T parameter for the scheduler.}

\item{\code{getToken()}}{:

Returns the access token.}

\item{\code{getUrls()}}{:

Returns the URLs.}

\item{\code{getWsUrl()}}{:

Returns the web sevices URL.}

\item{\code{getXmlNs()}}{:

Returns the XML namespace.}

\item{\code{setBaseUrl(base.url)}}{:

Sets the base URL.}

\item{\code{setSchedulerNParam(n)}}{:

Sets the N parameter for the scheduler.}

\item{\code{setSchedulerTParam(t)}}{:

Sets the T parameter for the scheduler.}

\item{\code{setToken(token)}}{:

Sets the access token.}

\item{\code{setWsUrl(ws.url)}}{:

Sets the web sevices URL.}

\item{\code{setXmlNs(xml.ns)}}{:

Sets the XML namespace.}
}}

\seealso{
\code{\link{BiodbDbsInfo}}, \code{\link{BiodbConn}}.
}
