% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbDbInfo.R
\docType{class}
\name{BiodbDbInfo-class}
\alias{BiodbDbInfo-class}
\alias{BiodbDbInfo}
\title{A class for describing the characteristics of a database.}
\description{
This class is used by \code{\link{BiodbDbsInfo}} for storing database
characteristics, and returning them through the \code{get()} method.
This class inherits from \code{\link{BiodbConnBase}}.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::BiodbMain()

# Get a BiodbDbInfo object for a database:
mybiodb$getDbsInfo()$get('comp.csv.file')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Parent class \code{\link{BiodbDbsInfo}} and super class
\code{\link{BiodbConnBase}}.
}
