% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbDbInfo.R
\docType{class}
\name{BiodbDbInfo-class}
\alias{BiodbDbInfo-class}
\alias{BiodbDbInfo}
\title{A class for describing the characteristics of a database.}
\description{
This class is used by \code{\link{BiodbDbsInfo}} for storing database characteristics.
}

\seealso{
\code{\link{BiodbDbsInfo}}.
}
