% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbDbsInfo.R
\docType{class}
\name{BiodbDbsInfo-class}
\alias{BiodbDbsInfo-class}
\alias{BiodbDbsInfo}
\title{A class for describing the available databases.}
\description{
A class for describing the available databases.
}

\seealso{
\code{\link{Biodb}}, \code{\link{BiodbDbInfo}}.
}
