% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbDownloadable.R
\docType{class}
\name{BiodbDownloadable-class}
\alias{BiodbDownloadable-class}
\alias{BiodbDownloadable}
\title{An abstract class (more like an interface) to model a remote database that
can be downloaded locally.}
\description{
The class must be inherited by any remote database class that allows download
of its whole content. The hidden/private method \code{.doDownload()} must be
implemented by the database class.
}
\section{Methods}{

\describe{
\item{\code{download()}}{:

Downloads the database content locally.
    
Returned value: None.
    }

\item{\code{getDownloadPath()}}{:

Gets the path where the downloaded content is written.
    
Returned value: The path where the downloaded database is written.
    }

\item{\code{isDownloaded()}}{:

Tests if the database has been downloaded.
    
Returned value: TRUE if the database content has already been downloaded.
    }

\item{\code{isExtracted()}}{:

Tests if the downloaded database has been extracted (in case the
    database needs extraction).
    
Returned value: TRUE if the downloaded database content has been
    extracted, FALSE otherwise.
    }
}}

\examples{
# Creating a connector class for a downloadable database:
FooDownloadableDb <- methods::setRefClass('FooDownloadableDb',
    contains=c('BiodbRemotedbConn', 'BiodbDownloadable'),

methods=list(

   # ... other methods ...

   # BiodbDownloadable methods
   .doDownload=function() {
       # Download the database
   },

   .doExtractDownload=function() {
       # Extract data from the downloaded file(s)
   }
))

}
\seealso{
\code{\link{BiodbRemotedbConn}}.
}
