% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbEditable.R
\docType{class}
\name{BiodbEditable-class}
\alias{BiodbEditable-class}
\alias{BiodbEditable}
\title{An interface to model an editable database.}
\description{
A database class that implements this interface allows the addition of new
entries, the removal of entries and the modification of entries. If you want
your modifications to be persistent, the database class must also be writable
(see interface \code{BiodbWritable}), you must call the method \code{write}
on the connector.
}
\section{Methods}{

\describe{
\item{\code{addNewEntry(entry)}}{:

Adds a new entry to the database. The passed entry must have been
    previously created from scratch using BiodbFactory::createNewEntry() or
    cloned from an existing entry using BiodbEntry::clone().
    
entry: The new entry to add. It must be a valid BiodbEntry object.
    
Returned value: None.
    }

\item{\code{allowEditing()}}{:

Allows editing for this database.
    
Returned value: None.
    }

\item{\code{disallowEditing()}}{:

Disallows editing for this database.
    
Returned value: None.
    }

\item{\code{editingIsAllowed()}}{:

Tests if editing is allowed.
    
Returned value: TRUE if editing is allowed for this database, FALSE
    otherwise.
    }

\item{\code{setEditingAllowed(allow)}}{:

Allow or disallow editing for this database.
    
allow: A logical value.
    
Returned value: None.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::BiodbMain()

# Create an empty MASS SQLite database
mydb <- mybiodb$getFactory()$createConn('mass.sqlite')

# Create new entry object
entry <- mybiodb$getFactory()$createNewEntry('mass.sqlite')
entry$setFieldValue('accession', '0')
entry$setFieldValue('name', 'Some Entry')

# Add the new entry
mydb$allowEditing()
mydb$addNewEntry(entry)

# Terminate instance.
mybiodb$terminate()
mybiodb <- NULL

}
\seealso{
\code{\link{BiodbConn}} and \code{\link{BiodbWritable}}.
}
