% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbEntryFields.R
\docType{class}
\name{BiodbEntryFields-class}
\alias{BiodbEntryFields-class}
\alias{BiodbEntryFields}
\title{A class for handling description of all entry fields.}
\arguments{
\item{database}{The name of a database.}
}
\description{
The unique instance of this class is handle by the \code{\link{Biodb}} class
and accessed through the \code{getEntryFields()} method.
}
\section{Methods}{

\describe{
\item{\code{checkIsDefined(name)}}{Throws an error if any name does not correspond to a defined field.
    
name: A character vector of names or aliases to test.}

\item{\code{get(name)}}{Returns the BiodbEntryField instance associated with name.}

\item{\code{getDatabaseIdField(database)}}{Returns the name of the field handling identifiers (i.e.: accession
numbers) for this database.}

\item{\code{getFieldNames(type = NULL)}}{Returns the main names of all fields.
    
type: If set, returns only the field names corresponding to this type.}

\item{\code{getRealName(name)}}{If name is an alias, returns the main name of the field. If name is not
found neither in aliases nor in real names, an error is thrown.}

\item{\code{isAlias(name)}}{Returns TRUE if name is an alias of a field.
    
name: A character vector of names or aliases to test.
    
Returned value: A logical vector.}

\item{\code{isDefined(name)}}{Returns TRUE if name corresponds to a defined field.
    
name: A character vector of names or aliases to test.
    
Returned value: A logical vector.}

\item{\code{loadFieldsFile(file)}}{Load entry fields information file, and defines the new fields.
The parameter file must point to a valid YAML file.}
}}

\examples{
# Getting information about the accession field:
mybiodb <- biodb::Biodb()
entry.field <- mybiodb$getEntryFields()$get('accession')

# Test if a name is an alias of a field
mybiodb$getEntryFields()$isAlias('genesymbols')

# Test if a name is associated with a defined field
mybiodb$getEntryFields()$isDefined('chebi.id')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{Biodb}}, \code{\link{BiodbEntryField}}.
}
