% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbEntryFields.R
\docType{class}
\name{BiodbEntryFields-class}
\alias{BiodbEntryFields-class}
\alias{BiodbEntryFields}
\title{A class for handing description of all entry fields.}
\description{
The unique instance of this class is handle by the \code{\link{Biodb}} class and accessed through the \code{getEntryFields()} method.
}

\examples{
# Getting information about the accession field:
mybiodb <- biodb::Biodb()
entry.field <- mybiodb$getEntryFields()$get('accession')

}
\seealso{
\code{\link{BiodbEntryField}}.
}
