% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbErrorReporter.R
\docType{class}
\name{BiodbErrorReporter-class}
\alias{BiodbErrorReporter-class}
\alias{BiodbErrorReporter}
\title{A class for logging error messages.}
\description{
This class is not meant to be used directly. It is automatically instantiate inside biodb constructor in order to report errors.
}

\seealso{
\code{\link{Biodb}}, \code{\link{BiodbObserver}}.
}
