% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbFactory.R
\docType{class}
\name{BiodbFactory-class}
\alias{BiodbFactory-class}
\alias{BiodbFactory}
\title{A class for constructing biodb objects.}
\arguments{
\item{content}{The content (as character vector) of one or more database entries.}

\item{db.class}{The type of a database. The list of types can be obtained from the class \code{\link{BiodbDbsInfo}}.}

\item{conn.id}{The identifier of a database connector.}

\item{drop}{If set to \code{TRUE} and the list of entries contains only one element, then returns this element instead of the list. If set to \code{FALSE}, then returns always a list.}

\item{id}{A character vector containing database entry IDs (accession numbers).}

\item{token}{A security access token for the database. Some database require such a token for all or some of their webservices. Usually you obtain the token through your account on the database website.}

\item{url}{An URL to the database for which to create a connection. Each database connector is configured with a default URL, but some allow you to change it.}
}
\description{
This class is responsible for the creation of database connectors and database entries. You must go through the single instance of this class to create and get connectors, as well as instantiate entries. To get the single instance of this class, call the \code{getFactory()} method of class \code{Biodb}.
}
\section{Methods}{

\describe{
\item{\code{connExists(conn.id)}}{:

Returns TRUE if a connector with this ID exists.}

\item{\code{createConn(db.class, url = NULL, token = NA_character_,
  fail.if.exists = TRUE, conn.id = NULL, cache.id = NULL)}}{:

Create a connection to a database.}

\item{\code{createNewEntry(db.class)}}{:

Create a new entry from scratch. This entry is not stored in cache.}

\item{\code{deleteAllCacheEntries(conn.id)}}{:

Delete all entries of a connector from the cache.}

\item{\code{deleteAllConnectors()}}{:

Delete all connectors.}

\item{\code{deleteConn(conn.id = NULL, db.class = NULL)}}{:

Delete existing connectors.}

\item{\code{getAllCacheEntries(conn.id)}}{:

Get all entries of a connector from the cache.}

\item{\code{getAllConnectors()}}{:

Returns a list of all created connectors.}

\item{\code{getConn(conn.id)}}{:

Get the instance of connector database corresponding to the submitted ID or database class.}

\item{\code{getEntry(conn.id, id, drop = TRUE)}}{:

Create database entry objects from IDs (accession numbers), for the specified connector.}
}}

\examples{
# Create a Biodb instance with default settings:
mybiodb <- biodb::Biodb()

# Obtain the factory instance:
factory <- mybiodb$getFactory()

# Create a connection:
conn <- factory$createConn('chebi')

# Get a database entry:
entry <- factory$getEntry('chebi', id = '2528')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{Biodb}}, \code{\link{BiodbConn}}, \code{\link{BiodbEntry}}.
}
