% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbFactory.R
\docType{class}
\name{BiodbFactory-class}
\alias{BiodbFactory}
\alias{BiodbFactory-class}
\title{A class for constructing biodb objects.}
\description{
A class for constructing biodb objects.
}
\section{Methods}{

\describe{
\item{\code{createConn(class, url = NA_character_, token = NA_character_)}}{Create connection to databases useful for metabolomics.}

\item{\code{getConn(class)}}{Get connection to a database.}

\item{\code{getEntry(class, id, drop = TRUE)}}{Create Entry from a database by id.}
}}

