% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbLogger.R
\docType{class}
\name{BiodbLogger-class}
\alias{BiodbLogger-class}
\alias{BiodbLogger}
\title{A class for logging biodb messages either to standard stream or into a file.}
\description{
This class implements a logger for the biodb package. When creating an
instance of this class, you can choose to either log to a standard stream
(standard error or standard output) or to a file. See section Fields for a
list of the constructor's parameters.
}
\section{Fields}{

\describe{
\item{\code{file}}{The file path, file connection or standard stream to which you
want to send biodb messages.}

\item{\code{mode}}{For a file, the mode in which you want to open the file. 'w' for
writing (if a file already exists, it will be erased), and 'a' for
appending (if a file already exists, logs will be appended to it).}
}}


\examples{
# Create a file logger
logger <- biodb::biodbLogger(file="myfile.log")

# Create a biodb instance
mybiodb <- biodb::Biodb()

# Add the logger to the list of observers
mybiodb$addObservers(logger)

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{Biodb}}, \code{\link{BiodbObserver}}.
}
