% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbLogger.R
\docType{class}
\name{BiodbLogger-class}
\alias{BiodbLogger-class}
\alias{BiodbLogger}
\title{A class for logging biodb messages either to standard stream or into a file.}
\description{
This class implements a logger for the biodb package. When creating an
instance of this class, you can choose to either log to a standard stream
(standard error or standard output) or to a file. See section Fields for a
list of the constructor's parameters.
}
\details{
The constructor accepts the following arguments:

file: The file path, file connection or standard stream to which you
      want to send biodb messages.

mode: For a file, the mode in which you want to open the file. 'w' for
      writing (if a file already exists, it will be erased), and 'a' for
      appending (if a file already exists, logs will be appended to it).

close.file: If set to TRUE, and `file` is a file path or a file connection,
            then the file connection will be closed when calling
            `terminate()`. If it is set to FALSE, the connection will not be
            closed.
}
\section{Methods}{

\describe{
\item{\code{msg(type = "info", msg, class = NA_character_, method = NA_character_,
  lvl = 1)}}{:

Sends a message to this observer. The message will be accepted and
    handled only if the level is lower or equal to the current level for the
    message type. You can check current level for a type by calling getLevel()
    and set the level with setLevel().
    
type: The message type. It must be one of: 'info', 'debug',
    'caution', 'warning', 'error'.
    
msg: The text message to send.
    
class: The class of the object that called this message method.
    
method: The method that called this message method.
    
lvl: The level of the message.
    
Returned value: None.
    }

\item{\code{terminate()}}{:

Terminates the instance. This method will be called
    automatically by the Biodb instance when you call Biodb::terminate().
    
Returned value: None.
    }
}}

\examples{
# Create a file logger
logger <- biodb::BiodbLogger(file="myfile.log")

# Create a biodb instance
mybiodb <- biodb::Biodb()

# Add the logger to the list of observers
mybiodb$addObservers(logger)

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{Biodb}}, \code{\link{BiodbObserver}}.
}
