% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbObserver.R
\docType{class}
\name{BiodbObserver-class}
\alias{BiodbObserver-class}
\alias{BiodbObserver}
\title{The mother abstract class of all observer classes.}
\description{
The mother abstract class of all observer classes.
}

