% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbRect.R
\docType{class}
\name{BiodbRect-class}
\alias{BiodbRect-class}
\alias{BiodbRect}
\title{A class for representing a rectangle.}
\description{
This class represents a rectangle, used for graphical
representation.
}
\section{Fields}{

\describe{
\item{\code{left}}{Coordinate of left border.}

\item{\code{right}}{Coordinate of right border.}

\item{\code{top}}{Coordinate of top border.}

\item{\code{bottom}}{Coordinate of bottom border.}
}}


\examples{
# Create a rectangle instance
r <- BiodbRect(left=10, top=10, bottom=20, right=30, color='yellow')

# Draw a rectangle on current image
\dontrun{
r$draw()
}

}
\seealso{
\code{\link{BiodbShape}}, \code{\link{BiodbCircle}}.
}
