% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbRemotedbConn.R
\docType{class}
\name{BiodbRemotedbConn-class}
\alias{BiodbRemotedbConn-class}
\alias{BiodbRemotedbConn}
\title{The mother class of all remote database connectors.}
\arguments{
\item{concatenate}{If set to \code{TRUE}, then try to build as few URLs as
possible, sending requests with several identifiers at once.}

\item{entry.id}{The identifiers (e.g.: accession numbers) as a
\code{character vector} of the database entries.}

\item{max.length}{The maximum length of the URLs to return, in number of
characters.}
}
\description{
This is the super class of remote database connectors. It thus defines
methods related to remote connection, like the definition of a token, and URL
definitions. As with \code{\link{BiodbConn}} class, you won't need to use the
constructor. Nevertheless we provide in the Fields section information about
the constructor parameters, for eventual developers.
}
\section{Methods}{

\describe{
\item{\code{getEntryContentFromDb(entry.id)}}{Get the content of an entry from the database.}

\item{\code{getEntryContentRequest(entry.id, concatenate = TRUE, max.length = 0)}}{Get the URL to use in order to get the contents of the specified entries.}

\item{\code{getEntryImageUrl(entry.id)}}{Get the URL to a picture of the entry (e.g.: a picture of the molecule in
case of a compound entry).}

\item{\code{getEntryPageUrl(entry.id)}}{Get the URL to the page of the entry on the database web site.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get connector
conn <- mybiodb$getFactory()$createConn('chemspider')

# Get the picture URL of an entry
picture.url <- conn$getEntryImageUrl('2')

# Get the page URL of an entry
page.url <- conn$getEntryPageUrl('2')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbConn}}, \code{\link{BiodbRequestScheduler}}.
}
