% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbRequestScheduler.R
\name{BiodbRequestScheduler}
\alias{BiodbRequestScheduler}
\title{Class for handling requests.}
\description{
Class for handling requests.

Class for handling requests.
}
\details{
This class handles GET and POST requests, as well as file downloading. Each
remote database connection instance creates an instance of
\code{BiodbRequestScheduler} for handling database connection. A timer is
used to schedule connections, and avoid sending too much requests to the
database. This class is not meant to be used directly by the library user.
See section Fields for a list of the constructor's parameters.
}
\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Get the scheduler
sched <- mybiodb$getRequestScheduler()

# Create a request object
u <- 'https://www.ebi.ac.uk/webservices/chebi/2.0/test/getCompleteEntity'
url <- BiodbUrl$new(url=u)
url$setParam('chebiId', 15440)
request <- BiodbRequest$new(method='get', url=url)

# Send request
sched$sendRequest(request)

# Terminate instance.
mybiodb$terminate()
mybiodb <- NULL

}
\seealso{
\code{\link{BiodbRequestSchedulerRule}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiodbRequestScheduler$new()}}
\item \href{#method-sendSoapRequest}{\code{BiodbRequestScheduler$sendSoapRequest()}}
\item \href{#method-sendRequest}{\code{BiodbRequestScheduler$sendRequest()}}
\item \href{#method-downloadFile}{\code{BiodbRequestScheduler$downloadFile()}}
\item \href{#method-notifyConnUrlsUpdated}{\code{BiodbRequestScheduler$notifyConnUrlsUpdated()}}
\item \href{#method-notifyConnSchedulerFrequencyUpdated}{\code{BiodbRequestScheduler$notifyConnSchedulerFrequencyUpdated()}}
\item \href{#method-getUrlString}{\code{BiodbRequestScheduler$getUrlString()}}
\item \href{#method-getUrl}{\code{BiodbRequestScheduler$getUrl()}}
\item \href{#method-findRule}{\code{BiodbRequestScheduler$findRule()}}
\item \href{#method-getConnectorRules}{\code{BiodbRequestScheduler$getConnectorRules()}}
\item \href{#method-registerConnector}{\code{BiodbRequestScheduler$registerConnector()}}
\item \href{#method-unregisterConnector}{\code{BiodbRequestScheduler$unregisterConnector()}}
\item \href{#method-getAllRules}{\code{BiodbRequestScheduler$getAllRules()}}
\item \href{#method-clone}{\code{BiodbRequestScheduler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$new(bdb)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendSoapRequest"></a>}}
\if{latex}{\out{\hypertarget{method-sendSoapRequest}{}}}
\subsection{Method \code{sendSoapRequest()}}{
Sends a SOAP request to a URL. Returns the string result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$sendSoapRequest(
  url,
  soap.request,
  soap.action = NA_character_,
  encoding = integer()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL to access, as a character string.}

\item{\code{soap.request}}{The XML SOAP request to send, as a character string.}

\item{\code{soap.action}}{The SOAP action to contact, as a character string.}

\item{\code{encoding}}{The encoding to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The results returned by the contacted server, as a single
string value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendRequest"></a>}}
\if{latex}{\out{\hypertarget{method-sendRequest}{}}}
\subsection{Method \code{sendRequest()}}{
Sends a request, and returns content result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$sendRequest(request, cache.read = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request}}{A BiodbRequest instance.}

\item{\code{cache.read}}{If set to TRUE, the cache system will be used. In case the
same request has already been already run and its results saved into the
cache, then the request is not run again, the targeted server not contacted,
and the results are directly loaded from the cache system.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The results returned by the contacted server, as a single
string value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadFile"></a>}}
\if{latex}{\out{\hypertarget{method-downloadFile}{}}}
\subsection{Method \code{downloadFile()}}{
Downloads the content of a URL and save it into the specified
destination file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$downloadFile(url, dest.file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL to access, as a BiodbUrl object.}

\item{\code{dest.file}}{A path to a destination file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-notifyConnUrlsUpdated"></a>}}
\if{latex}{\out{\hypertarget{method-notifyConnUrlsUpdated}{}}}
\subsection{Method \code{notifyConnUrlsUpdated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$notifyConnUrlsUpdated(conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-notifyConnSchedulerFrequencyUpdated"></a>}}
\if{latex}{\out{\hypertarget{method-notifyConnSchedulerFrequencyUpdated}{}}}
\subsection{Method \code{notifyConnSchedulerFrequencyUpdated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$notifyConnSchedulerFrequencyUpdated(conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUrlString"></a>}}
\if{latex}{\out{\hypertarget{method-getUrlString}{}}}
\subsection{Method \code{getUrlString()}}{
Build a URL string, using a base URL and parameters to be passed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$getUrlString(url, params = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getUrl}{}}}
\subsection{Method \code{getUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$getUrl(
  url,
  params = list(),
  method = c("get", "post"),
  header = character(),
  body = character(),
  encoding = integer()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findRule"></a>}}
\if{latex}{\out{\hypertarget{method-findRule}{}}}
\subsection{Method \code{findRule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$findRule(url, create = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConnectorRules"></a>}}
\if{latex}{\out{\hypertarget{method-getConnectorRules}{}}}
\subsection{Method \code{getConnectorRules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$getConnectorRules(conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerConnector"></a>}}
\if{latex}{\out{\hypertarget{method-registerConnector}{}}}
\subsection{Method \code{registerConnector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$registerConnector(conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unregisterConnector"></a>}}
\if{latex}{\out{\hypertarget{method-unregisterConnector}{}}}
\subsection{Method \code{unregisterConnector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$unregisterConnector(conn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllRules"></a>}}
\if{latex}{\out{\hypertarget{method-getAllRules}{}}}
\subsection{Method \code{getAllRules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$getAllRules()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbRequestScheduler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
