% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbRequestSchedulerRule.R
\docType{class}
\name{BiodbRequestSchedulerRule-class}
\alias{BiodbRequestSchedulerRule-class}
\alias{BiodbRequestSchedulerRule}
\title{Scheduler rule class.}
\description{
This class represents a rule for the request scheduler.
}
\section{Fields}{

\describe{
\item{\code{n}}{The number of connections allowed for each t seconds.}

\item{\code{t}}{The number of seconds during which n connections are allowed.}
}}


\seealso{
\code{\link{BiodbRequestScheduler}}.
}
