% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbShape.R
\docType{class}
\name{BiodbShape-class}
\alias{BiodbShape-class}
\alias{BiodbShape}
\title{A class for representing a shape.}
\description{
This abstract class represents a shape, used for graphical representation.
}
\section{Fields}{

\describe{
\item{\code{label}}{A text label to associate with the shape.}

\item{\code{color}}{A color, as a character string.}
}}

\section{Methods}{

\describe{
\item{\code{getColor()}}{Return the color associated with this shape.}

\item{\code{getLabel()}}{Return the label associated with this shape.}

\item{\code{getRgbColor(alpha = 255)}}{Return the RGB color associated with this shape.}
}}

\seealso{
\code{\link{BiodbRect}}, \code{\link{BiodbCircle}}.
}
