% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlBinaryOp.R
\name{BiodbSqlBinaryOp}
\alias{BiodbSqlBinaryOp}
\title{This class represents an SQL binary operator.}
\description{
This class represents an SQL binary operator.

This class represents an SQL binary operator.
}
\section{Super class}{
\code{\link[biodb:BiodbSqlExpr]{biodb::BiodbSqlExpr}} -> \code{BiodbSqlBinaryOp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BiodbSqlBinaryOp-new}{\code{BiodbSqlBinaryOp$new()}}
\item \href{#method-BiodbSqlBinaryOp-toString}{\code{BiodbSqlBinaryOp$toString()}}
\item \href{#method-BiodbSqlBinaryOp-clone}{\code{BiodbSqlBinaryOp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlBinaryOp-new"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlBinaryOp-new}{}}}
\subsection{Method \code{new()}}{
Initializer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlBinaryOp$new(lexpr, op, rexpr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lexpr}}{A BiodbSqlExpr instance for the left part.}

\item{\code{op}}{The binary operator, as a string.}

\item{\code{rexpr}}{A BiodbSqlExpr instance for the right part.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlBinaryOp-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlBinaryOp-toString}{}}}
\subsection{Method \code{toString()}}{
Converts into a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlBinaryOp$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string containing the SQL expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlBinaryOp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlBinaryOp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlBinaryOp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
