% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbSqlList.R
\name{BiodbSqlList}
\alias{BiodbSqlList}
\title{This class represents an SQL list.}
\description{
This class represents an SQL list.

This class represents an SQL list.
}
\section{Super class}{
\code{\link[biodb:BiodbSqlExpr]{biodb::BiodbSqlExpr}} -> \code{BiodbSqlList}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BiodbSqlList-new}{\code{BiodbSqlList$new()}}
\item \href{#method-BiodbSqlList-toString}{\code{BiodbSqlList$toString()}}
\item \href{#method-BiodbSqlList-clone}{\code{BiodbSqlList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlList-new"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlList-new}{}}}
\subsection{Method \code{new()}}{
Initializer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlList$new(values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{The values of the list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlList-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlList-toString}{}}}
\subsection{Method \code{toString()}}{
Converts into a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlList$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string containing the SQL expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiodbSqlList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BiodbSqlList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiodbSqlList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
