% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\docType{class}
\name{BiodbTestMsgAck-class}
\alias{BiodbTestMsgAck-class}
\alias{BiodbTestMsgAck}
\title{A class for acknowledging messages during tests.}
\description{
This observer is used to call a testthat::expect_*() method each time a
message is received. This is used when running tests on Travis-CI, so Travis
does not stop tests because no change is detected in output.
}
\section{Methods}{

\describe{
\item{\code{notifyProgress(what, index, total)}}{:

Notify about the progress of an action.
    
what: A short description of the action.
    
index: A positive integer number indicating the progress.
    
total: The maximum for "index". When reached, the action is completed.
    
Returned value: None.
    }
}}

\examples{
# To use the acknowledger, set ack=TRUE when creating the Biodb test
# instance:
biodb <- createBiodbTestInstance(ack=TRUE)

# Terminate the Biodb instance
biodb$terminate()

}
