% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\docType{class}
\name{BiodbTestMsgAck-class}
\alias{BiodbTestMsgAck-class}
\alias{BiodbTestMsgAck}
\title{A class for acknowledging messages during tests.}
\description{
This observer is used to call a testthat::expect_*() method each time a
message is received. This is used when running tests on Travis-CI, so Travis
does not stop tests because no change is detected in output.
}
\section{Methods}{

\describe{
\item{\code{msg(type = "info", msg, class = NA_character_, method = NA_character_,
  lvl = 1)}}{:

Sends a message to this observer. The message will be accepted and
    handled only if the level is lower or equal to the current level for the
    message type. You can check current level for a type by calling getLevel()
    and set the level with setLevel().
    
type: The message type. It must be one of: 'info', 'debug',
    'caution', 'warning', 'error'.
    
msg: The text message to send.
    
class: The class of the object that called this message method.
    
method: The method that called this message method.
    
lvl: The level of the message.
    
Returned value: None.
    }

\item{\code{progress(type = "info", msg, index, first, total = NA_integer_, lvl = 1L,
  laptime = 10L)}}{:

Sends a progress message to this observer.
    
type: The message type. It must be one of: 'info', 'debug',
    'caution', 'warning', 'error'.
    
msg: The text message to send.
    
index: The index in the progression, as an integer or numeric number.
    
total: The total to achieve in the progression, as an integer or numeric
    number. Optional.
    
lvl: The level of the message.
    
laptime: The time between two progress messages, in seconds.
    
Returned value: None.
    }
}}

