% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbUrl.R
\docType{class}
\name{BiodbUrl-class}
\alias{BiodbUrl-class}
\alias{BiodbUrl}
\title{Class URL.}
\description{
This class represents a URL object that can be used in requests.
}
\details{
The following arguments are accepted by the constructor:

url: The URL to access, as a character string.

params: The list of parameters to use with the URL.
}
\section{Methods}{

\describe{
\item{\code{getDomain()}}{:

Gets the domain.
    
Returned value: None.
    }

\item{\code{setParam(key, value)}}{:

Sets a parameter.
    
key:
    
value:
    
Returned value: None.
    }

\item{\code{setUrl(url)}}{:

Sets the base URL string.
    
url: The base URL string.
    
Returned value: None.
    }

\item{\code{toString(encode = TRUE)}}{:

Gets the URL as a string representation.
    
encode: If set to TRUE, then encodes the URL.
    
Returned value: The URL as a string, with all parameters and values set.
    }
}}

\examples{
# Create a URL object
u <- c("https://www.uniprot.org", "uniprot")
p <- c(query="reviewed:yes+AND+organism:9606",
       columns='id,entry name,protein names',
       format="tab")
url <- BiodbUrl(url=u, params=p)
url$toString()

}
\seealso{
\code{\link{BiodbRequestScheduler}}, \code{\link{BiodbRequest}}.
}
