% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodbUrl.R
\docType{class}
\name{BiodbUrl-class}
\alias{BiodbUrl-class}
\alias{BiodbUrl}
\title{Class URL.}
\arguments{
\item{url}{The URL to access, as a character string.}

\item{params}{The list of parameters to use with the URL.}
}
\description{
This class represents a URL object that can be used in requests.
}

\examples{
# Create a URL object
u <- c("https://www.uniprot.org", "uniprot")
p <- c(query="reviewed:yes+AND+organism:9606",
       columns='id,entry name,protein names',
       format="tab")
url <- BiodbUrl(url=u, params=p)
url$toString()

}
\seealso{
\code{\link{BiodbRequestScheduler}}, \code{\link{BiodbRequest}}.
}
