% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_biodb.R
\name{checkDeprecatedCacheFolders}
\alias{checkDeprecatedCacheFolders}
\title{Check deprecated default cache folders.}
\usage{
checkDeprecatedCacheFolders()
}
\description{
Check deprecated default cache folders.
}
