% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemspiderConn.R
\docType{class}
\name{ChemspiderConn-class}
\alias{ChemspiderConn-class}
\alias{ChemspiderConn}
\title{The connector class to ChemSpider database.}
\arguments{
\item{range}{}
}
\description{
This is a concrete connector class. It must never be instantiated directly, but instead be instantiated through the factory \code{\link{BiodbFactory}}. Only specific methods are described here. See super classes for the description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryContent(entry.id)}}{:

Get the content of an entry.}

\item{\code{getEntryIds(max.results = NA_integer_)}}{:

Get entry identifiers from the database.}

\item{\code{getEntryImageUrl(entry.id)}}{:

Get the URL to a picture of the entry (e.g.: a picture of the molecule in case of a compound entry).}

\item{\code{getEntryPageUrl(entry.id)}}{:

Get the URL to the page of the entry on the database web site.}

\item{\code{ws.SearchByMass2(mass = NA, range = NA)}}{Direct query to the database for searching for compounds by mass. See http://www.chemspider.com/MassSpecAPI.asmx?op=SearchByMass2 for details.}
}}

\seealso{
\code{\link{BiodbFactory}}, \code{\link{RemotedbConn}}, \code{\link{CompounddbConn}}.
}
