% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompounddbConn.R
\docType{class}
\name{CompounddbConn-class}
\alias{CompounddbConn-class}
\alias{CompounddbConn}
\title{An interface for all compound databases.}
\arguments{
\item{name}{The name of a compound.}

\item{mass}{The searched molecular mass.}

\item{mass.tol}{The tolerance on the molecular mass.}

\item{mass.tol.unit}{The unit used for molecular mass tolerance. Either 'plain' or 'ppm'.}

\item{max.results}{The maximum number of matches wanted.}
}
\description{
This interface must be inherited by all compound databases. It declares method headers specific to compound databases.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get the connector of a compound database
uniprot <- mybiodb$getFactory()$createConn('uniprot')

# Search for compounds
uniprot$searchCompound(name = 'prion protein', max.results = 10)

}
\seealso{
\code{\link{BiodbConn}}.
}
