% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{doesRCurlRequestUrlExist}
\alias{doesRCurlRequestUrlExist}
\title{Test if a URL is valid according to RCurl}
\usage{
doesRCurlRequestUrlExist(request, useragent = NULL)
}
\arguments{
\item{request}{A BiodbRequest object, from which the URL will be gotten.}

\item{useragent}{The user agent identification.}
}
\value{
Returns TRUE if the URL
}
\description{
Test if a URL is valid according to RCurl
}
