% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpasyEnzymeConn.R
\docType{class}
\name{ExpasyEnzymeConn-class}
\alias{ExpasyEnzymeConn-class}
\alias{ExpasyEnzymeConn}
\title{The connector class to Expasy Enzyme database.}
\arguments{
\item{name}{The name to search for.}

\item{comment}{The comment to search for.}
}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{\link{BiodbFactory}}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }

\item{\code{wsEnzymeByComment(comment, retfmt = c("plain", "request", "parsed", "ids"))}}{Calls enzyme-bycomment web service and returns the HTML result. See
http://enzyme.expasy.org/enzyme-bycomment.html.}

\item{\code{wsEnzymeByName(name, retfmt = c("plain", "request", "parsed", "ids"))}}{Calls enzyme-byname web service and returns the HTML result. See
http://enzyme.expasy.org/enzyme-byname.html.}
}}

