% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtPackageFile.R
\name{ExtPackageFile}
\alias{ExtPackageFile}
\title{Extension package file class.}
\description{
A class for generating the package.R file for a biodb extension.
}
\details{
This class generates the package.R file, writing a reference to the
generated skeleton vignette, and possibly including directives for C++ code.
}
\examples{
# Generate a new package:
pkgFolder <- file.path(tempfile(), 'biodbFoo')
dir.create(pkgFolder, recursive=TRUE)
biodb::ExtPackageFile$new(path=pkgFolder, dbName='foo.db')$generate()

}
\section{Super classes}{
\code{biodb::ExtGenerator} -> \code{biodb::ExtFileGenerator} -> \code{ExtPackageFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExtPackageFile-new}{\code{ExtPackageFile$new()}}
\item \href{#method-ExtPackageFile-clone}{\code{ExtPackageFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biodb" data-topic="ExtGenerator" data-id="generate"><a href='../../biodb/html/ExtGenerator.html#method-ExtGenerator-generate'><code>biodb::ExtGenerator$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="ExtGenerator" data-id="upgrade"><a href='../../biodb/html/ExtGenerator.html#method-ExtGenerator-upgrade'><code>biodb::ExtGenerator$upgrade()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExtPackageFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExtPackageFile-new}{}}}
\subsection{Method \code{new()}}{
Initializer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtPackageFile$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See the constructor of ExtFileGenerator for the parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExtPackageFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExtPackageFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtPackageFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
