% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtReadme.R
\name{ExtReadme}
\alias{ExtReadme}
\title{Extension README file class}
\description{
A class for generating a README file for a new extension package.
}
\details{
Write a README file inside package directory, using a template file.
}
\examples{
# Generate a new package:
pkgFolder <- file.path(tempfile(), 'biodbFoo')
dir.create(pkgFolder, recursive=TRUE)
biodb::ExtReadme$new(path=pkgFolder, dbName='foo.db',
                     dbTitle='Foo database')$generate()

}
\section{Super classes}{
\code{biodb::ExtGenerator} -> \code{biodb::ExtFileGenerator} -> \code{ExtReadme}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExtReadme-new}{\code{ExtReadme$new()}}
\item \href{#method-ExtReadme-clone}{\code{ExtReadme$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biodb" data-topic="ExtGenerator" data-id="generate"><a href='../../biodb/html/ExtGenerator.html#method-ExtGenerator-generate'><code>biodb::ExtGenerator$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodb" data-topic="ExtGenerator" data-id="upgrade"><a href='../../biodb/html/ExtGenerator.html#method-ExtGenerator-upgrade'><code>biodb::ExtGenerator$upgrade()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExtReadme-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExtReadme-new}{}}}
\subsection{Method \code{new()}}{
Initializer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtReadme$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See the constructor of ExtFileGenerator for the parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExtReadme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExtReadme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExtReadme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
