% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{getBaseUrlRequestResult}
\alias{getBaseUrlRequestResult}
\title{Get URL request result using base::url().}
\usage{
getBaseUrlRequestResult(request)
}
\arguments{
\item{request}{A BiodbRequest object.}
}
\value{
A RequestResult object.
}
\description{
Get URL request result using base::url().
}
