% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{getRCurlContent}
\alias{getRCurlContent}
\title{Get URL content using RCurl::getURL().}
\usage{
getRCurlContent(
  u,
  opts = NULL,
  enc = integer(),
  header.fct = NULL,
  ssl.verifypeer = TRUE,
  method = c("get", "post"),
  binary = FALSE
)
}
\arguments{
\item{u}{The URL as a character value.}

\item{opts}{A valid RCurl options object.}

\item{enc}{The encoding.}

\item{header.fct}{An RCurl header gatherer function.}

\item{ssl.verifypeer}{Set to TRUE to enable SSL verify peer.}

\item{method}{The HTTP method to use, either 'get' or 'post'.}

\item{binary}{Set to TRUE if the content to be retrieved is binary.}
}
\value{
The URL content as a character single value.
}
\description{
Get URL content using RCurl::getURL().
}
