% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_url.R
\name{getRCurlRequestResult}
\alias{getRCurlRequestResult}
\title{Get URL request result using RCurcl::getURL().}
\usage{
getRCurlRequestResult(request, opts, ssl.verifypeer = TRUE)
}
\arguments{
\item{request}{A BiodbRequest object.}

\item{opts}{A valid RCurl options object.}
}
\value{
A RequestResult object.
}
\description{
Get URL request result using RCurcl::getURL().
}
