% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmdbMetabolitesEntry.R
\docType{class}
\name{HmdbMetabolitesEntry-class}
\alias{HmdbMetabolitesEntry-class}
\alias{HmdbMetabolitesEntry}
\title{HMDB Metabolites entry class.}
\description{
This is the entry class for the HMDB Metabolites database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('hmdb.metabolites')

# Get an entry
\dontrun{
e <- conn$getEntry('HMDB0000001')
}

# Terminate instance.
mybiodb$terminate()

}
