% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggConn.R
\docType{class}
\name{KeggConn-class}
\alias{KeggConn}
\alias{KeggConn-class}
\title{The mother abstract class of all KEGG connection classes.}
\description{
The mother abstract class of all KEGG connection classes.
}

