% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggConn.R
\docType{class}
\name{KeggConn-class}
\alias{KeggConn-class}
\alias{KeggConn}
\title{The connector abstract class to KEGG databases.}
\arguments{
\item{query}{The query to send to the database web service.}
}
\description{
This is the mother class of all KEGG connectors. It defines code common to all KEGG connectors.
}
\section{Methods}{

\describe{
\item{\code{getEntryIds(max.results = NA_integer_)}}{:

Get entry identifiers from the database.}

\item{\code{getEntryPageUrl(entry.id)}}{:

Get the URL to the page of the entry on the database web site.}

\item{\code{ws.find(query)}}{:

Search for entries. See http://www.kegg.jp/kegg/docs/keggapi.html for details.}

\item{\code{ws.find.df(...)}}{:

Calls ws.find() and returns a data frame.}

\item{\code{ws.find.ids(...)}}{:

Calls ws.find() but only for getting IDs. Returns the IDs as a character vector.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector to a KEGG database
conn <- mybiodb$getFactory()$createConn('kegg.compound')

# Search for an entry
conn$ws.find.df('NADPH')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{RemotedbConn}}.
}
