% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggConn.R
\docType{class}
\name{KeggConn-class}
\alias{KeggConn-class}
\alias{KeggConn}
\title{The connector abstract class to KEGG databases.}
\arguments{
\item{query}{The query to send to the database web service.}
}
\description{
This is the mother class of all KEGG connectors. It defines code common to
all KEGG connectors.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{Get the URL to the page of the entry on the database web site.}

\item{\code{wsFind(query, retfmt = c("plain", "request", "parsed", "ids"))}}{Search for entries. See http://www.kegg.jp/kegg/docs/keggapi.html for
details.}

\item{\code{wsList(retfmt = c("plain", "request", "ids"))}}{Get list of entry IDs. See http://www.kegg.jp/kegg/docs/keggapi.html for
details.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector to a KEGG database
conn <- mybiodb$getFactory()$createConn('kegg.compound')

# Search for an entry
conn$wsFind('NADPH', retfmt='parsed')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{BiodbRemotedbConn}},
\code{\link{BiodbSearchable}}.
}
