% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggEnzymeConn.R
\docType{class}
\name{KeggEnzymeConn-class}
\alias{KeggEnzymeConn-class}
\alias{KeggEnzymeConn}
\title{The connector class to KEGG Enzyme database.}
\description{
This is a concrete connector class. It must never be instantiated directly, but instead be instantiated through the factory \code{\link{BiodbFactory}}. Only specific methods are described here. See super classes for the description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryImageUrl(entry.id)}}{:

Get the URL to a picture of the entry (e.g.: a picture of the molecule in case of a compound entry).}

\item{\code{getPathwayIds(id, org)}}{Get organism pathways. Given a vector of KEGG Enzyme IDs and a KEGG organism code, this method retrieves KEGG pathways of this organism in which the enzymes are involved. It returns a vector of KEGG pathway IDs.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get connector
conn <- mybiodb$getFactory()$createConn('kegg.enzyme')

# Get pathway IDs related to enzymes
pathway.ids = conn$getPathwayIds(c('1.2.1.3', '3.7.1.3'), org = 'mmu')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{KeggConn}}.
}
