% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggEnzymeConn.R
\docType{class}
\name{KeggEnzymeConn-class}
\alias{KeggEnzymeConn-class}
\alias{KeggEnzymeConn}
\title{The connector class to KEGG Enzyme database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{\link{BiodbFactory}}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getPathwayIds(id, org)}}{:

Gets organism pathways.  This method retrieves KEGG pathways of the
    specified organism in which the enzymes are involved.
    
id: A character vector of KEGG Compound IDs.
    
org: The organism in which to search for pathways, as a KEGG organism code
    (3-4 letters code, like 'hsa', 'mmu', ...). See
    https://www.genome.jp/kegg/catalog/org_list.html for a complete list of KEGG
    organism codes.
    
Returned value: A vector of KEGG pathway IDs.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get connector
conn <- mybiodb$getFactory()$createConn('kegg.enzyme')

# Get pathway IDs related to enzymes
pathway.ids=conn$getPathwayIds(c('1.2.1.3', '3.7.1.3'), org='mmu')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{KeggConn}}.
}
