% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidmapsStructureEntry.R
\docType{class}
\name{LipidmapsStructureEntry-class}
\alias{LipidmapsStructureEntry-class}
\alias{LipidmapsStructureEntry}
\title{Lipidmaps Structure entry class.}
\description{
Lipidmaps Structure entry class.
}

