% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_framework.R
\name{listTestRefEntries}
\alias{listTestRefEntries}
\title{List test reference entries.}
\usage{
listTestRefEntries(conn.id)
}
\arguments{
\item{conn.id}{A valid Biodb connector ID.}
}
\value{
A list of entry IDs.
}
\description{
Lists the reference entries in the test folder for a specified connector.
The test reference files must be in `<pkg>/tests/testthat/res/` folder and
their names must match `entry-<database_name>-<entry_accession>.json` (e.g.:
`entry-chebi-15440.json`).
}
\examples{
# List IDs of test reference entries for ChEBI:
biodb::listTestRefEntries('chebi')

}
