% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassdbConn.R
\docType{class}
\name{MassdbConn-class}
\alias{MassdbConn-class}
\alias{MassdbConn}
\title{The mother class of all Mass spectra databases.}
\arguments{
\item{max.results}{The maximum of elements returned by a method.}

\item{min.rel.int}{The minimum relative intensity.}

\item{ms.level}{The MS level to which you want to restrict your search. \code{0} means that you want to serach in all levels.}

\item{ms.mode}{The MS mode. Set it to either \code{BIODB.MSMODE.NEG} or \code{BIODB.MSMODE.POS}.}

\item{mz}{An M/Z value.}

\item{mz.max}{The maximum allowed for searched M/Z values.}

\item{mz.min}{The minimum allowed for searched M/Z values.}

\item{plain.tol}{The M/Z tolerance in M/Z unit. Thus the searched mz must satisfy the folloing double inequality: mz - plain.tol <= mz <= mz + plain.tol.}

\item{precursor}{If set to \code{TRUE}, then restrict the search to precursor peaks.}

\item{tol}{An M/Z tolerance, whose unit is not defined.}

\item{tol.unit}{The unit of the M/Z tolerance. Set it to either \code{BIODB.MZTOLUNIT.PPM} or \code{BIODB.MZTOLUNIT.PLAIN}.}
}
\description{
All Mass spectra databases inherit from this class.
}
\section{Methods}{

\describe{
\item{\code{getMzValues(ms.mode = NA_character_, max.results = NA_integer_,
  precursor = FALSE, ms.level = 0)}}{Get a list of M/Z values contained inside the database.}

\item{\code{searchMzRange(mz.min, mz.max, min.rel.int = NA_real_,
  ms.mode = NA_character_, max.results = NA_integer_, precursor = FALSE,
  ms.level = 0)}}{Find spectra in the given M/Z range. Returns a list of spectra IDs.}

\item{\code{searchMzTol(mz, tol, tol.unit = BIODB.MZTOLUNIT.PLAIN,
  min.rel.int = NA_real_, ms.mode = NA_character_,
  max.results = NA_integer_, precursor = FALSE, ms.level = 0)}}{Find spectra containg a peak around the given M/Z value. Returns a list of spectra IDs.}
}}

