% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSqliteConn.R
\docType{class}
\name{MassSqliteConn-class}
\alias{MassSqliteConn-class}
\alias{MassSqliteConn}
\title{Class for handling a Mass spectrometry database in SQLite format.}
\description{
Class for handling a Mass spectrometry database in SQLite format.
}

