% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSqliteConn.R
\docType{class}
\name{MassSqliteConn-class}
\alias{MassSqliteConn-class}
\alias{MassSqliteConn}
\title{Class for handling a Mass spectrometry database in SQLite format.}
\description{
Class for handling a Mass spectrometry database in SQLite format.
}
\section{Methods}{

\describe{
\item{\code{getChromCol(ids = NULL)}}{:

Get a list of chromatographic columns contained in this database. You can filter on specific entries using the ids parameter. The returned value is a data.frame with two columns : one for the ID 'id' and another one for the title 'title'.}

\item{\code{getEntryContentFromDb(entry.id)}}{:

Get the content of an entry from the database.}
}}

