% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSqliteConn.R
\docType{class}
\name{MassSqliteConn-class}
\alias{MassSqliteConn-class}
\alias{MassSqliteConn}
\title{Class for handling a Mass spectrometry database in SQLite format.}
\description{
This is the connector class for a MASS SQLite database.
}
\section{Methods}{

\describe{
\item{\code{getChromCol(ids = NULL)}}{:

Gets a list of chromatographic columns contained in this database.
    
ids: A character vector of entry identifiers (i.e.: accession numbers).
    Used to restrict the set of entries on which to run the algorithm.
    
Returned value : A data.frame with two columns, one for the ID 'id' and
    another one for the title 'title'.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get path to LCMS database example file
lcmsdb <- system.file("extdata", "massbank_extract.sqlite", package="biodb")

# Create a connector
conn <- mybiodb$getFactory()$createConn('mass.sqlite', url=lcmsdb)

# Get an entry
e <- conn$getEntry('34.pos.col12.0.78')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super classes \code{\link{BiodbMassdbConn}} and
\code{\link{SqliteConn}}.
}
