% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSqliteConn.R
\docType{class}
\name{MassSqliteConn-class}
\alias{MassSqliteConn-class}
\alias{MassSqliteConn}
\title{Class for handling a Mass spectrometry database in SQLite format.}
\description{
Class for handling a Mass spectrometry database in SQLite format.
}
\section{Methods}{

\describe{
\item{\code{getChromCol(ids = NULL)}}{:

Gets a list of chromatographic columns contained in this database.
    
ids: A character vector of entry identifiers (i.e.: accession numbers).
    Used to restrict the set of entries on which to run the algorithm.
    
Returned value : A data.frame with two columns, one for the ID 'id' and
    another one for the title 'title'.
    }

\item{\code{getEntryContentFromDb(entry.id)}}{:

Get the contents of entries directly from the database. A direct request or 
    an access to the database will be made in order to retrieve the contents. No
    access to the biodb cache system will be made.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A character vector, the same size of entry.id, with
    contents of the requested entries. An NA value will be set for the content
    of each entry for which the retrieval failed.
    }
}}

