% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirbaseMatureConn.R
\docType{class}
\name{MirbaseMatureConn-class}
\alias{MirbaseMatureConn-class}
\alias{MirbaseMatureConn}
\title{Mirbase Mature connector class.}
\description{
This is the connector class for Mirbase Mature database.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('mirbase.mature')

# Get an entry
e <- conn$getEntry('MIMAT0000433')

# Terminate instance.
mybiodb$terminate()

}
