% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiCcdsConn.R
\docType{class}
\name{NcbiCcdsConn-class}
\alias{NcbiCcdsConn-class}
\alias{NcbiCcdsConn}
\title{NCBI CCDS connector class.}
\description{
NCBI CCDS connector class.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }
}}

