% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiConn.R
\docType{class}
\name{NcbiConn-class}
\alias{NcbiConn-class}
\alias{NcbiConn}
\title{The mother abstract class of all NCBI connection classes.}
\description{
The mother abstract class of all NCBI connection classes.
}

