% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemCompConn.R
\docType{class}
\name{NcbiPubchemCompConn-class}
\alias{NcbiPubchemCompConn-class}
\alias{NcbiPubchemCompConn}
\title{NCBI PubChem Compound connector class.}
\description{
This is the connector class for a NCBI PubChen Compound database.
}
\section{Methods}{

\describe{
\item{\code{searchCompound(name = NULL, mass = NULL, mass.field = NULL,
  mass.tol = 0.01, mass.tol.unit = "plain", max.results = NA_integer_)}}{Search for compounds by name and/or by mass. At least one of name or mass
    must be set.
    
name: The name of a compound to search for.
    
mass: The searched mass.
    
mass.field: For searching by mass, you must indicate a mass field to use
    ('monoisotopic.mass', 'molecular.mass', 'average.mass' or 'nominal.mass').
    
mass.tol: The tolerance value on the molecular mass.
    
mass.tol.unit: The type of mass tolerance. Either 'plain' or 'ppm'.
    
max.results: The maximum number of matches to return.
    
Returned value: A character vector of entry IDs.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.pubchem.comp')

# Get an entry
e <- conn$getEntry('2')

# Terminate instance.
mybiodb$terminate()

}
