% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemCompEntry.R
\docType{class}
\name{NcbiPubchemCompEntry-class}
\alias{NcbiPubchemCompEntry-class}
\alias{NcbiPubchemCompEntry}
\title{NCBI PubChem Compound entry class.}
\description{
This is the entry class for a NCBI PubChen Compound database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('ncbi.pubchem.comp')

# Get an entry
e <- conn$getEntry('2')

# Terminate instance.
mybiodb$terminate()

}
