% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemSubstConn.R
\docType{class}
\name{NcbiPubchemSubstConn-class}
\alias{NcbiPubchemSubstConn-class}
\alias{NcbiPubchemSubstConn}
\title{NCBI PubChem Substance connector class.}
\description{
NCBI PubChem Substance connector class.
}

