% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiPubchemSubstEntry.R
\docType{class}
\name{NcbiPubchemSubstEntry-class}
\alias{NcbiPubchemSubstEntry-class}
\alias{NcbiPubchemSubstEntry}
\title{NCBI PubChem Substance entry class.}
\description{
NCBI PubChem Substance entry class.
}

