% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakforestCompoundEntry.R
\docType{class}
\name{PeakforestCompoundEntry-class}
\alias{PeakforestCompoundEntry-class}
\alias{PeakforestCompoundEntry}
\title{PeakForest Compound entry class.}
\description{
This is the entry class for PeakForest Compound database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Create a connector
conn <- mybiodb$getFactory()$createConn('peakforest.compound')

# Get an entry
e <- conn$getEntry('1839')

# Terminate instance.
mybiodb$terminate()

}
