% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_tests.R
\name{runGenericTests}
\alias{runGenericTests}
\title{Run generic tests.}
\usage{
runGenericTests(conn)
}
\arguments{
\item{conn}{A valid biodb connector.}
}
\description{
This function must be used in tests on all connector classes, before any
specific tests.
}
