% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestRefEntries.R
\name{TestRefEntries}
\alias{TestRefEntries}
\title{A class for accessing the test reference entries.}
\description{
A class for accessing the test reference entries.

A class for accessing the test reference entries.
}
\details{
The test reference entries are stored as JSON files inside
\code{inst/testref} folder of each extension package.
}
\examples{
# Creates an instance
refEntries <- TestRefEntries$new('comp.sqlite', pkgName='biodb')

# Gets identifiers of all reference entries
refEntries$getAllIds()

# Gets a data frame with the content of the reference entries
refEntries$getAllRefEntriesDf()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TestRefEntries$new()}}
\item \href{#method-getAllIds}{\code{TestRefEntries$getAllIds()}}
\item \href{#method-getRealEntries}{\code{TestRefEntries$getRealEntries()}}
\item \href{#method-getAllRefEntriesDf}{\code{TestRefEntries$getAllRefEntriesDf()}}
\item \href{#method-clone}{\code{TestRefEntries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
New instance initializer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRefEntries$new(db.class, pkgName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db.class}}{Identifier of the database.}

\item{\code{pkgName}}{Name of the package in which are stored the reference
entry files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllIds"></a>}}
\if{latex}{\out{\hypertarget{method-getAllIds}{}}}
\subsection{Method \code{getAllIds()}}{
Retrieve all identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRefEntries$getAllIds(limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{The maximum number of identifiers to return.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector containing the IDs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRealEntries"></a>}}
\if{latex}{\out{\hypertarget{method-getRealEntries}{}}}
\subsection{Method \code{getRealEntries()}}{
Retrieve all real entries from database corresponding to the reference
entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRefEntries$getRealEntries(bdb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bdb}}{A valid BiodbMain instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing Entry instances.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllRefEntriesDf"></a>}}
\if{latex}{\out{\hypertarget{method-getAllRefEntriesDf}{}}}
\subsection{Method \code{getAllRefEntriesDf()}}{
Load all reference entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRefEntries$getAllRefEntriesDf()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame containing all the reference entries with their values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRefEntries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
