% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotConn.R
\docType{class}
\name{UniprotConn-class}
\alias{UniprotConn-class}
\alias{UniprotConn}
\title{The connector class to Uniprot database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{\link{BiodbFactory}}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryPageUrl(entry.id)}}{:

Gets the URL to the page of the entry on the database web site.
    
entry.id: A character vector with the IDs of entries to retrieve.
    
Returned value: A list of BiodbUrl objects, the same length as `entry.id`.
    }

\item{\code{searchCompound(name = NULL, mass = NULL, mass.field = NULL,
  mass.tol = 0.01, mass.tol.unit = "plain", max.results = NA_integer_)}}{Search for compounds by name and/or by mass. At least one of name or mass
    must be set.
    
name: The name of a compound to search for.
    
mass: The searched mass.
    
mass.field: For searching by mass, you must indicate a mass field to use
    ('monoisotopic.mass', 'molecular.mass', 'average.mass' or 'nominal.mass').
    
mass.tol: The tolerance value on the molecular mass.
    
mass.tol.unit: The type of mass tolerance. Either 'plain' or 'ppm'.
    
max.results: The maximum number of matches to return.
    
Returned value: A character vector of entry IDs.}

\item{\code{wsQuery(query = "", columns = NULL, format = NULL, limit = NULL,
  retfmt = c("plain", "parsed", "ids", "request"))}}{:

Calls query to the database for searching for compounds.
    See http://www.uniprot.org/help/api_queries for details.
    
query: The query to send to the database.
    
columns: The field columns to retrieve from the database (e.g.: 'id',
    'entry name', 'pathway', 'organism', 'sequence', etc).
    
format: The return format (e.g.: 'tab').
    
limit: The maximum number of entries to return.
    
retfmt: Use to set the format of the returned value. 'plain' will return
    the raw results from the server, as a character value. 'parsed' will return
    the parsed results, as a JSON object. 'request' will return a BiodbRequest
    object representing the request as it would have been sent. 'ids' will
    return a character vector containing the IDs of the matching entries.
    
Returned value: Depending on `retfmt` parameter.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get Uniprot connector
uniprot <- mybiodb$getFactory()$createConn('uniprot')

# Access web service query
result <- uniprot$wsQuery(query='name:"prion protein"',
                           columns=c('id', 'entry name'),
                           format='txt', limit=10)

# Terminate instance.
mybiodb$terminate()

}
\seealso{
Super classes \code{\link{BiodbRemotedbConn}} and
\code{\link{BiodbCompounddbConn}}.
}
