% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotConn.R
\docType{class}
\name{UniprotConn-class}
\alias{UniprotConn-class}
\alias{UniprotConn}
\title{The connector class to Uniprot database.}
\arguments{
\item{columns}{The field columns to retrieve from the database.}

\item{format}{The return format.}

\item{query}{The query to send to the database.}
}
\description{
This is a concrete connector class. It must never be instantiated directly, but instead be instantiated through the factory \code{\link{BiodbFactory}}. Only specific methods are described here. See super classes for the description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryContent(entry.id)}}{:

Get the content of an entry.}

\item{\code{getEntryIds(max.results = NA_integer_)}}{:

Get entry identifiers from the database.}

\item{\code{ws.query(query = "", columns = NA, format = NA, limit = NA)}}{Direct query to the database for searching for compounds. See http://www.uniprot.org/help/api_queries for details.}

\item{\code{ws.query.ids(...)}}{Calls ws.query() but only for getting IDs. Returns the IDs as a character vector.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::Biodb()

# Get Uniprot connector
uniprot <- mybiodb$getFactory()$createConn('uniprot')

# Access web service query
result <- uniprot$ws.query(query = 'name:"prion protein"', columns = c('id', 'entry name'), format = 'txt', limit = 10)

}
\seealso{
\code{\link{BiodbFactory}}, \code{\link{RemotedbConn}}, \code{\link{CompounddbConn}}.
}
